% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aesthetics}
\alias{aesthetics}
\title{\code{ggplot2} Manual Scale Values}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 5 columns.
}
\usage{
aesthetics
}
\description{
\code{\link[tibble:tibble]{tibble::tibble()}} of manual scale values for \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\details{
\describe{
\item{species}{Taxonomic annotation matching value in \link{Specimen}}
\item{color}{Color value character scalar}
\item{shape}{Shape value numeric scalar}
\item{text-color}{Descriptive text of color aesthetic}
\item{text-shape}{Descriptive text of shape aesthetic}
}
}
\keyword{datasets}
