% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{annotation_legend}
\alias{annotation_legend}
\title{Extract Legend of \link{ggplot} Aesthetics from Specimen Annotations}
\usage{
annotation_legend(
  specimen_tbl,
  aesthetic_id,
  ncol = 2,
  legend_title,
  aesthetic_labels = NULL,
  theme.params = list()
)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{aesthetic_id}{Variable to set color and shape aesthetics of jitters.}

\item{ncol}{Number of legend columns.}

\item{legend_title}{Character vector of length one to set the ggplot
legend title.}

\item{aesthetic_labels}{Named vector of labels for manual scale values.}

\item{theme.params}{List of arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A \code{grob} with extracted plot legend of annotation aesthetics.
}
\description{
Extract Legend of \link{ggplot} Aesthetics from Specimen Annotations
}
\examples{
specimen_tbl <- herbarium_specimens \%>\%
\dontrun{
  subset_coords(Longitude = c(-108, -105), Latitude = c(39, 42))
aesthetic_id <- "Taxon_a_posteriori"
legend_title <- "Reviewed Annotations"
aesthetic_labels <- NULL
theme.params <- list()
guide.params <- list()
}
}
\keyword{internal}
