% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_bayes.R
\name{bayes_ggtree}
\alias{bayes_ggtree}
\title{Build \code{ggtree} Plot}
\usage{
bayes_ggtree(joined_ggtree, id_column, ggtree_layout = "slanted")
}
\arguments{
\item{joined_ggtree}{Tibble of \code{ggtree} nodes with labels and scaled geoms
from DNA specimens output by \code{\link[=join_bayes]{join_bayes()}}.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}

\item{ggtree_layout}{\code{layout} argument passed to \code{\link[ggtree:ggtree]{ggtree::ggtree()}}.}
}
\value{
\code{ggtree} object built from \code{\link[=read_tree]{read_tree()}} and \code{\link[=join_bayes]{join_bayes()}} outputs.
}
\description{
Given output of \code{\link[=read_tree]{read_tree()}}, build a \code{\link[ggtree:ggtree]{ggtree::ggtree()}} plot with
single- and multi-taxa tip labels.
}
\examples{
joined_ggtree <-
  list.files(system.file("extdata/MrBayes", package = "thesis"),
    full.names = TRUE, pattern = "rITS-infile.nex.con.tre"
  ) \%>\%
  read_tree(tree_file = .) \%>\%
  join_bayes(tree_data = ., id_column = "prior_id")

# Build ggtre object from MrBayes data import.
bayes_ggtree(
  joined_ggtree = joined_ggtree,
  id_column = "prior_id"
)

}
\keyword{internal}
