% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\name{bayes_kable}
\alias{bayes_kable}
\title{Kable build for multi-specimen nodes.}
\usage{
bayes_kable(conserved_specimens, knitr_chunk, kable_caption, kable_scap = "")
}
\arguments{
\item{conserved_specimens}{Voucher tibble output by \code{\link[=conserved_vouchers]{conserved_vouchers()}}}

\item{knitr_chunk}{Character vector passed to \code{\link[knitr:kable]{knitr::kable()}} argument
\code{format} - one of "latex" or "html".  Defined by
\code{opts_knit$get("rmarkdown.pandoc.to")} in \code{index.Rmd} of thesis Bookdown.}

\item{kable_caption}{Character vector for caption passed to
\code{\link[knitr:kable]{knitr::kable()}} argument \code{caption}.}

\item{kable_scap}{Character vector for caption passed to
\code{\link[knitr:kable]{knitr::kable()}} argument \code{caption}.}
}
\value{
\code{kable} object with \code{kableExtra} formatting determined by
\code{knitr::opts_knit$get("rmarkdown.pandoc.to")}
}
\description{
For specimens with multi-taxa nodes (i.e., terminal tips), build
\code{kable} object with \code{kableExtra} formatting.
}
\examples{
list.files(system.file("extdata/MrBayes", package = "thesis"),
  full.names = TRUE, pattern = "rITS-infile.nex.con.tre"
) \%>\%
  conserved_vouchers(
    tree_file = .,
    id_column = "prior_id", id_name = "Species"
  ) \%>\%
  bayes_kable(
    conserved_specimens = .,
    kable_caption = "ggtree Specimens", knitr_chunk = "html"
  )

}
\keyword{internal}
