% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_bayes.R
\name{bayes_plot}
\alias{bayes_plot}
\title{Build ggtree Plot Grid}
\usage{
bayes_plot(
  joined_ggtree,
  id_column,
  scale_name,
  x_expand = 0.02,
  plot_x = 0.1,
  plot_y = 0.5,
  plot_width = 0.25,
  plot_height = 0.5,
  ...
)
}
\arguments{
\item{joined_ggtree}{Tibble of \code{ggtree} nodes with labels and scaled geoms
from DNA specimens output by \code{\link[=join_bayes]{join_bayes()}}.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}

\item{scale_name}{Character scalar to set color and shape legend name.}

\item{x_expand}{Numeric scalar passed to \code{\link[ggplot2:expand_limits]{ggplot2::expand_limits()}}.}

\item{plot_x}{Numeric scalar for plot x position.}

\item{plot_y}{Numeric scalar for plot y position.}

\item{plot_width}{Numeric scalar for plot width.}

\item{plot_height}{Numeric scale for plot height.}

\item{...}{Forwarding arguments to \code{\link[cowplot:draw_plot]{cowplot::draw_plot()}}.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} with \code{\link[ggtree:ggtree]{ggtree::ggtree()}} and legend inset.
}
\description{
Wrapper function to plot legend inset onto ggtree object.
}
\examples{
joined_ggtree <-
  list.files(system.file("extdata/MrBayes", package = "thesis"),
    full.names = TRUE, pattern = "rITS-infile.nex.con.tre"
  ) \%>\%
  read_tree(tree_file = .) \%>\%
  join_bayes(
    tree_data = ., id_column = "prior_id",
    scale_vector = c(4, 12)
  )

# Build ggtree plot grid.
bayes_plot(joined_ggtree,
  x_expand = 0.025, plot_y = 0.45,
  id_column = "prior_id", scale_name = "Priors"
)

}
\keyword{internal}
