% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_bayes.R
\name{bayes_themes}
\alias{bayes_themes}
\title{MrBayes \code{ggtree} Theme Layer}
\usage{
bayes_themes(joined_ggtree, id_column)
}
\arguments{
\item{joined_ggtree}{Tibble of \code{ggtree} nodes with labels and scaled geoms
from DNA specimens output by \code{\link[=join_bayes]{join_bayes()}}.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}
}
\value{
List of \code{ggplot} theme specifications for \code{ggtree} build.
}
\description{
Build list of \code{ggplot} theme options for MrBayes \code{ggtree} build.
}
\examples{
joined_ggtree <-
  list.files(system.file("extdata/MrBayes", package = "thesis"),
    full.names = TRUE, pattern = "rITS-infile.nex.con.tre"
  ) \%>\%
  read_tree(tree_file = .) \%>\%
  join_bayes(
    tree_data = ., id_column = "prior_id",
    scale_vector = c(4, 12)
  )

# Add theme layers to ggtree plot.
bayes_ggtree(joined_ggtree, id_column = "prior_id") +
  bayes_themes(joined_ggtree, id_column = "prior_id")

}
\keyword{internal}
