% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_beast.R
\name{beast_labels}
\alias{beast_labels}
\title{BEAST Specimen Labels}
\usage{
beast_labels(tree_data)
}
\arguments{
\item{tree_data}{Tibble output by \code{\link[=read_tree]{read_tree()}} with ggtree node and
label variables.}
}
\value{
\code{\link[ggtree:ggtree]{ggtree::ggtree()}} aesthetic layers for specimen labels.
}
\description{
Build specimen tip and node labels with posterior probability > 0.5.
Includes both prior and reviewed annotations.
}
\examples{
tree_data <-
  list.files(
    path = system.file("extdata/BEAST", package = "thesis"),
    pattern = "multi-locus.combined.mcc", full.names = TRUE
  ) \%>\%
  read_tree(tree_file = .)

ggtree::ggtree(tr = tree_data) +
  beast_labels(tree_data = tree_data)

}
\keyword{internal}
