% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_beast.R
\name{beast_legend_color}
\alias{beast_legend_color}
\title{Extract Color Legend}
\usage{
beast_legend_color(tree_data, ncol = 2)
}
\arguments{
\item{tree_data}{Tibble output by \code{\link[=read_tree]{read_tree()}} with ggtree node and
label variables.}

\item{ncol}{Number of columns for annotation legend guide.}
}
\value{
\code{gtable} object with specimen label legend.
}
\description{
Extract Color Legend
}
\keyword{internal}
