% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_beast.R
\name{beast_legend_probability}
\alias{beast_legend_probability}
\title{Extract Probability Legend}
\usage{
beast_legend_probability(tree_data)
}
\arguments{
\item{tree_data}{Tibble output by \code{\link[=read_tree]{read_tree()}} with ggtree node and
label variables.}
}
\value{
\code{gtable} object with posterior probability legend.
}
\description{
Extract Probability Legend
}
\keyword{internal}
