% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_beast.R
\name{beast_plot}
\alias{beast_plot}
\title{BEAST Plot Grid}
\usage{
beast_plot(tree_data, ggtree_layout = "circular")
}
\arguments{
\item{tree_data}{Tibble output by \code{\link[=read_tree]{read_tree()}} with ggtree node and
label variables.}

\item{ggtree_layout}{\code{layout} argument passed to \code{\link[ggtree:ggtree]{ggtree::ggtree()}}.}
}
\value{
\code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} of BEAST results with posterior probabilities,
specimen labels, and theme-layered \code{\link[ggtree:ggtree]{ggtree::ggtree()}} object.
}
\description{
Build a grid plot by \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} from BEAST tree data.
Includes a tree aesthetic with posterior probabilities and color
aesthetic for layering prior and reviewed annotations.
}
\examples{
tree_data <-
  list.files(
    path = system.file("extdata/BEAST", package = "thesis"),
    pattern = "multi-locus.combined.mcc", full.names = TRUE
  ) \%>\%
  read_tree(tree_file = .)

beast_plot(tree_data = tree_data)

}
\keyword{internal}
