% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_beast.R
\name{beast_posterior}
\alias{beast_posterior}
\title{BEAST Probability Tree Aesthetic}
\usage{
beast_posterior()
}
\value{
\code{\link[ggtree:ggtree]{ggtree::ggtree()}} aesthetic layers for posterior probabilities.
}
\description{
Build \code{\link[ggtree:geom_tree]{ggtree::geom_tree()}} layer with posterior probabilities from
BEAST results.
}
\examples{
list.files(
  path = system.file("extdata/BEAST", package = "thesis"),
  pattern = "multi-locus.combined.mcc", full.names = TRUE
) \%>\%
  read_tree(tree_file = .) \%>\%
  ggtree::ggtree(tr = .) +
  beast_posterior()

}
\keyword{internal}
