% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_beast.R
\name{beast_theme}
\alias{beast_theme}
\title{BEAST \code{ggtree} Theme}
\usage{
beast_theme(tree_data)
}
\arguments{
\item{tree_data}{Tibble output by \code{\link[=read_tree]{read_tree()}} with ggtree node and
label variables.}
}
\value{
\code{\link[ggtree:ggtree]{ggtree::ggtree()}} aesthetic layers for tree themes.
}
\description{
Build tree layers to specify manual color and shape values and set legend
theme. Legend includes all values from prior and reviewed annotations in
\code{tree_data}.
}
\examples{
tree_data <-
  list.files(
    path = system.file("extdata/BEAST", package = "thesis"),
    pattern = "multi-locus.combined.mcc", full.names = TRUE
  ) \%>\%
  read_tree(tree_file = .)

ggtree::ggtree(tr = tree_data) +
  beast_labels(tree_data = tree_data) +
  beast_theme(tree_data = tree_data)

}
\keyword{internal}
