% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bem}
\alias{bem}
\title{sp object of Barnsley East and Mexborough parliamentary constituency}
\format{The @data contains the following variables:
\describe{
  \item{name}{Name of parliamentary constituency}
  \item{label}{Label of parliamentary constituency}
}}
\source{
\url{https://borders.ukdataservice.ac.uk/ukborders/easy_download/prebuilt/}
}
\usage{
bem
}
\description{
A spatialPolygonsDataFrame object with outline of Barnsley East and
Mexborough Westminster Parliamentary Constituency based on 2001 boundaries.
}
\keyword{datasets}
