% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-maps.R
\name{build_map}
\alias{build_map}
\title{Build map from layer functions}
\usage{
build_map(
  specimen_tbl,
  id_column,
  legend_title,
  map_base = c("base", "ggmap", "elevation"),
  ...
)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}

\item{legend_title}{Character vector of length one to set the ggplot
legend title.}

\item{map_base}{Character vector to determine base ggplot layer.}

\item{...}{Forward arguments to: \link{layer_ggmap}, \link{layer_elevation},
\link{layer_borders}, \link{layer_specimens}, \link{layer_themes}, and
\link{spl_legend}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Assemble ggplot from various layer functions. Plot includes a base layer of
bare, ggmap, or elevation rasters, county and state borders, specimens, and
theme layers. The \link{cowplot} package is used to create a plot grid of the
ggplot build.
}
\examples{
specimen_tbl <- spp_co_front_range
id_column <- "prior_id"
legend_title <- "Priors"
map_base <- "base"
map_base <- "ggmap"
map_base <- "elevation"

}
\keyword{internal}
