% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_funcs.R
\name{calc_perc_error}
\alias{calc_perc_error}
\title{calc_perc_error}
\usage{
calc_perc_error(area_sim_values, area_cen_values, zone_pops)
}
\arguments{
\item{area_sim_values}{A vector of simulated values for each zone}

\item{area_cen_values}{A vector of known (usually census) values for each
zone}

\item{zone_pops}{A vector of known populations for each zone}
}
\value{
perc_error
}
\description{
Calculates the percentage error of a simulation
}
