% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_funcs.R
\name{calc_sae}
\alias{calc_sae}
\title{calc_sae}
\usage{
calc_sae(tae, sim_pops)
}
\arguments{
\item{tae}{A vector of total absolute error for each zone, usually
calculated with calc_tae()}

\item{sim_pops}{A vector of simulated zone populations}
}
\value{
sae
}
\description{
Calculate the standardised absolute error (SAE) of a simulation.
Takes TAE as an argument, usually calculated using calc_tae()
}
