% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_funcs.R
\name{calc_sei}
\alias{calc_sei}
\title{calc_sei}
\usage{
calc_sei(area_sim_values, area_cen_values)
}
\arguments{
\item{area_sim_values}{A vector of simulated values for each zone}

\item{area_cen_values}{A vector of known (usually census) values for each
zone}
}
\value{
sei
}
\description{
Calculate the standard error around the identity (SEI) of a simulation
}
