% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_funcs.R
\name{calc_tae}
\alias{calc_tae}
\title{calc_tae}
\usage{
calc_tae(con_pops, sim_pops)
}
\arguments{
\item{con_pops}{Known/actual population usually taken from the constraints.
Should be a vector of numeric populations, one for each zone}

\item{sim_pops}{Simulated populations. Should be a vector of numeric
populations, one for each zone}
}
\value{
tae
}
\description{
Calculate the total absolute error (TAE) of a simulation
}
