% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize String Character}
\usage{
capitalize(character_vector)
}
\arguments{
\item{character_vector}{Character vector to capitalize.}
}
\description{
Given an input character vector, capitalize the first element character.
}
\examples{
traits <- c("oblong", "elliptic", "linear")
capitalize(character_vector = traits)

}
\seealso{
Other text: 
\code{\link{italicize}()}
}
\concept{text}
\keyword{internal}
