% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{care_homes}
\alias{care_homes}
\title{sp object to plot Doncaster care homes}
\format{The @data contains the following variables:
\describe{
  \item{name}{Care home name}
  \item{postcode}{Postcode of care home}
  \item{type}{Type of care home: residential, nursing, or both}
}}
\source{
Care Quality Commission,
(\url{http://www.cqc.org.uk/content/using-cqc-data#directory}),
geocoding performed with Google Geocoding API
}
\usage{
care_homes
}
\description{
A spatialPointsDataFrame object of Doncaster care homes, including nursing
homes and residential homes (residential homes have no nursing facilities),
in British National Grid projection (EPSG:27700).
}
\keyword{datasets}
