% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_functions.R
\name{check_logit}
\alias{check_logit}
\title{check_logit}
\usage{
check_logit(model)
}
\arguments{
\item{model}{a logit model object to test}
}
\value{
A list containing two data frames ('ind' and 'over'):
\itemize{
  \item \code{ind}
    \itemize{
      \item \code{predictor} Predictor variable
      \item \code{beta} Beta coefficient
      \item \code{lower_ci} Lower odds ratio confidence interval (95\%)
      \item \code{odds_ratio} Odds ratio
      \item \code{upper_ratio} Upper odds ratio confidence interval (95\%)
    }
  \item \code{over}
    \itemize{
      \item \code{diff_deviance} the difference between the null deviance
      and model deviance
      \item \code{diff_df} the difference in degrees of freedom
      \item \code{chisq_prob} the chi-squared probability of the deviance statistic
      \item \code{cox_snell} Cox and Snell pseudo R-squared
      \item \code{nagelkerke} Nagelkerke pseudo R-squared
      \item \code{hosmer} Hosmer-Lemeshow pseudo R-squared
    }
}
}
\description{
Check logistic regression statistics and assumptions
}
