% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R, R/rel_weight.R
\name{compute_rel_weight}
\alias{compute_rel_weight}
\title{This function calculates the relative weight of each sector for each day 
and display the result in a dataframe}
\usage{
compute_rel_weight()

compute_rel_weight()
}
\description{
This function calculates the relative weight of each sector for each day 
and display the result in a dataframe

This function calculates the relative weight of each sector for each day 
and display the result in a dataframe
}

