% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\name{conserved_vouchers}
\alias{conserved_vouchers}
\title{Multi-taxa Node Specimens}
\usage{
conserved_vouchers(tree_file, id_column, id_name)
}
\arguments{
\item{tree_file}{Nexus file of MrBayes results to read in by \code{treeio}.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}

\item{id_name}{Name to replace \code{id_column} variable in kable output.}
}
\value{
Tibble of specimens comprising multi-taxa nodes with identical
DNA sequences for MrBayes analysis.
}
\description{
Identify specimens with conserved DNA sequences yielding identical nodes
(i.e., terminal tips) for phylogenetic inference.
}
\examples{
list.files(system.file("extdata/MrBayes", package = "thesis"),
  full.names = TRUE, pattern = "rITS-infile.nex.con.tre"
) \%>\%
  conserved_vouchers(
    tree_file = .,
    id_column = "prior_id", id_name = "Species"
  )

}
\keyword{internal}
