% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-specimens.R
\name{count_specimens}
\alias{count_specimens}
\title{Count Specimens}
\usage{
count_specimens(spp_tibble)
}
\arguments{
\item{spp_tibble}{Specimen tibble to count unique specimen vouchers.}
}
\value{
Numeric scalar with the number of distinct specimen vouchers.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Tabulate the total number of distinct specimen collections by collector,
colleciton number and date while accounting for herbarium duplicates.
}
\examples{
count_specimens(spp_tibble = herbarium_specimens)

}
\keyword{internal}
