% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doctors}
\alias{doctors}
\title{sp object to plot Doncaster care homes}
\format{The @data contains the following variables:
\describe{
  \item{name}{Doctors surgery name}
  \item{postcode}{Postcode of doctor surgery}
  \item{type}{Type (used to filter from master cqc file)}
}}
\source{
Care Quality Commission,
(\url{http://www.cqc.org.uk/content/using-cqc-data#directory}),
geocoding performed with Google Geocoding API
}
\usage{
doctors
}
\description{
A spatialPointsDataFrame object of Doncaster care homes, including nursing
homes and residential homes (residential homes have no nursing facilities),
in British National Grid projection (EPSG:27700).
}
\keyword{datasets}
