% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{don_context}
\alias{don_context}
\title{A list of 4 sp objects containing geographical context for don_outline}
\format{An object of class \code{list} of length 4.}
\source{
Ordnance Survey Strategi
\url{https://www.ordnancesurvey.co.uk/business-and-government/products/strategi.html}
}
\usage{
don_context
}
\description{
The 4 shapefiles are:
}
\details{
\itemize{
  \item don_context[[1]]: mways - motorways - line
  \item don_context[[2]]: a_rds - 'A' roads - line
  \item don_context[[3]]: p_rds - Primary roads - line
  \item don_context[[4]]: urban - urban areas - polygons
}
}
\keyword{datasets}
