% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{don_msoa}
\alias{don_msoa}
\title{sp object to plot Doncaster MSOAs}
\format{The @data contains the following variables:
\describe{
  \item{code}{output area code}
  \item{pop_16_plus}{Population aged 16 and above at 2011 Census}
  \item{pop_65_plus}{Population aged 65 and above at 2011 Census}
  \item{llid_cen}{Number of people with a limiting long--term illness or disability taken from 2011 Census, either day--to--day activities limited a lot or a little.}
  \item{cen_llid_yes_p}{Proportion (0 <= x <= 1) llid_cen}
}}
\source{
Output area shapefile from \url{https://census.edina.ac.uk/easy_download_data.html?data=England_oa_2011}.
OA to MSOA lookup from ONS using geolookup package (\url{https://github.com/philmikejones/geolookup})
}
\usage{
don_msoa
}
\description{
A spatialPolygonsDataFrame object of Doncaster middle layer super output
areas LSOAs).
Simplified with rmapshaper::ms_simplify().
Subset from England and Wales output areas obtained from
\url{http://geoportal.statistics.gov.uk/datasets/}
}
\keyword{datasets}
