% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{don_oa}
\alias{don_oa}
\title{sp object to plot Doncaster output areas}
\format{The @data contains the following variables:
\describe{
  \item{code}{output area code}
  \item{spg_code}{Output area classification supergroup code}
  \item{spg_name}{Output area classification supergroup name}
  \item{g_code}{Output area classification group code}
  \item{g_name}{Output area classification group name}
  \item{sb_code}{Output area classification subgroup code}
  \item{sb_name}{Output area classification subgroup name}
  \item{pop_16_plus}{Population aged 16 and above at 2011 Census}
  \item{pop_65_plus}{Population aged 65 and above at 2011 Census}
  \item{yp_pop}{Population of young people (aged <18)}
  \item{prison_pop}{Prisoner population for that output area, NA if no prison (source: Census 2011 obtained from Nomisweb)}
  \item{llid_cen}{Number of people with a limiting long--term illness or disability taken from 2011 Census, either day--to--day activities limited a lot or a little.}
  \item{cen_llid_yes_p}{Proportion (0 <= x <= 1) llid_cen}
  \item{unem}{Number of economically active unemployed}
  \item{higher_man_1}{NS-SEC: higher managerial}
  \item{lower_man_2}{NS-SEC: lower managerial}
  \item{intermediate_3}{NS-SEC: intermediate occupations}
  \item{self_emp_4}{NS-SEC: small employers and own-account}
  \item{lower_sup_5}{NS-SEC: lower supervisory and technical}
  \item{semi_routine_6}{NS-SEC: Semi-routine occupations}
  \item{routine_7}{NS-SEC: routine occupations}
  \item{lt_unem_8}{NS-SEC: Never worked and long-term unemployed}
  \item{codecomm}{Community area code for joining}
  \item{namecomm}{Community area name for joining}
}}
\source{
\itemize{
  \item Output area shapefile from \url{https://census.edina.ac.uk/easy_download_data.html?data=England_oa_2011}.
  \item Output area classifications from \url{http://webarchive.nationalarchives.gov.uk/20160105160709/http://www.ons.gov.uk/ons/guide-method/geography/products/area-classifications/ns-area-classifications/ns-2011-area-classifications/datasets/index.html}.
  \item Prison locations obtained from geocoding postcodes.
  \item Economically active unemployed from 2011 Census
  \item NS-SEC from 2011 Census
}
}
\usage{
don_oa
}
\description{
A spatialPolygonsDataFrame object of Doncaster output areas.
Simplified with rmapshaper::ms_simplify().
Subset from England and Wales output areas obtained from
\url{http://geoportal.statistics.gov.uk/datasets/}
}
\keyword{datasets}
