% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{don_oa_cent}
\alias{don_oa_cent}
\title{sp object of Doncaster OA centroids}
\format{The @data contains the following variables:
\describe{
  \item{code}{output area code}
}}
\source{
\itemize{
  \item Output area shapefile from \url{https://census.edina.ac.uk/easy_download_data.html?data=England_oa_2011}
}
}
\usage{
don_oa_cent
}
\description{
A spatialPointsDataFrame object of Doncaster output area centroids.
}
\keyword{datasets}
