% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{don_outline}
\alias{don_outline}
\title{sp object of Doncaster local authority district (LAD)}
\format{The @data contains the following variables:
\describe{
  \item{code}{Local authority district (LAD) code}
  \item{name}{LAD name}
  \item{region}{Government office region}
  \item{supergroup}{Supergroup area classification of the LAD}
  \item{group}{Group area classification of the LAD}
  \item{subgroup}{Subgroup classification of the LAD}
}}
\source{
\itemize{
  \item Boundary data: \url{https://census.edina.ac.uk/easy_download_data.html?data=infuse_dist_lyr_2011}
  \item Area classification: \url{http://www.ons.gov.uk/ons/guide-method/geography/products/area-classifications/ns-area-classifications/ns-2011-area-classifications/datasets/2011-census-data.xls}
}
}
\usage{
don_outline
}
\description{
Doncaster local authority district (LADs) saved as a polygon spatial object,
including area classifications.
}
\keyword{datasets}
