% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-specimens.R
\name{filter_reviewed}
\alias{filter_reviewed}
\title{Filter Reviewed Specimens}
\usage{
filter_reviewed(specimen_tbl)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}
}
\value{
Tibble of specimens filtered by species of reviewed annotation.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Subset specimen tibble to exclude outgroups, leaving only species of interest.
}
\details{
Reviewed species include:
\itemize{
\item Physaria
\item P. acutifolia
\item P. brassicoides
\item P. condensata
\item P. didymocarpa subsp. didymocarpa
\item P. didymocarpa subsp. lanata
\item P. didymocarpa subsp. lyrata
\item P. dornii
\item P. eburniflora
\item P. integrifolia
\item P. 'medicinae'
\item P. vitulifera
}
}
\examples{
herbarium_specimens \%>\%
  filter_reviewed(specimen_tbl = .) \%>\%
  dplyr::select(Taxon_a_posteriori) \%>\%
  dplyr::add_count(Taxon_a_posteriori) \%>\%
  dplyr::distinct()

}
\keyword{internal}
