% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-specimens.R
\name{find_spp}
\alias{find_spp}
\title{Find specimens by Collector and Collection Number.}
\usage{
find_spp(specimen_tbl, collector = NULL, collection = NULL)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens built from
\code{vouchers.R} script.}

\item{collector}{Character scalar to filter specimen by collector name.}

\item{collection}{Numeric vector to filter specimens by collection number.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
find_spp(herbarium_specimens, collector = "Rollins")
find_spp(herbarium_specimens, collector = "Rollins", collection = 81337)

}
\keyword{internal}
