% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendtomatch.R, R/g_with_multassign.R
\name{g}
\alias{g}
\alias{g}
\title{extendToMatch}
\usage{
g(...)

g(...)
}
\arguments{
\item{...}{list}

\item{source}{}
}
\description{
Extends the one side to match the other side if you need to. 
inspired by http://stackoverflow.com/questions/7519790/assign-multiple-new-variables-on-lhs-in-a-single-line-in-r.

groups a list for use with the %=% operator. 
inspired by http://stackoverflow.com/questions/7519790/assign-multiple-new-variables-on-lhs-in-a-single-line-in-r.
}
\seealso{
%=%

%=%
}
\keyword{assign}
