% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gb_outline}
\alias{gb_outline}
\title{sp object to plot the outline of Great Britain}
\format{The @data contains the following variables:
\describe{
  \item{geo_code}{Geography code}
  \item{geo_label}{Name or label of the geography}
  \item{geo_labelw}{Name or label of the geography in Welsh}
  \item{label}{Geography code}
  \item{name}{Name of the geography}
}}
\source{
\url{https://census.edina.ac.uk/easy_download_data.html?data=infuse_gb_2011}
}
\usage{
gb_outline
}
\description{
A spatialPolygonsDataFrame object to plot an outline of Great Britain.
}
\keyword{datasets}
