% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{geocode}
\usage{
geocode(address, key, output_format = "json")
}
\arguments{
\item{address}{A human-readable address or postcode. Should not contain
special characters like spaces.}

\item{key}{Either a character string containing your API key or the location
of a .txt file containing the string on the first line only}

\item{output_format}{json or xml (currently only json is supported and is
the default)}
}
\value{
A data frame with two columns, lat and long
}
\description{
Geocode an address or postcode with Google Geocode API
Require an API key which can be obtained from:
https://developers.google.com/maps/documentation/geocoding/get-api-key
}
\details{
Geocoding converts an address or postcode to latitude/longitude coordinates

Address should not contain special characters like spaces (which should be
replaced with '+'). See https://developers.google.com/maps/documentation/geocoding/web-service-best-practices#BuildingURLs
}
