% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multassign lbunch.R
\name{\%=\%.lbunch}
\alias{\%=\%.lbunch}
\title{%=%.lbunch}
\usage{
\method{\%=\%}{lbunch}(l, r, ...)
}
\arguments{
\item{l}{left hand list}

\item{r}{right hand list}
}
\description{
groups the left hand side for use with the %=% operator. 
inspired by http://stackoverflow.com/questions/7519790/assign-multiple-new-variables-on-lhs-in-a-single-line-in-r.
}
\keyword{assign}
