% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multassign.R
\name{\%=\%}
\alias{\%=\%}
\title{%=%}
\usage{
l \%=\% r
}
\arguments{
\item{l}{left hand list}

\item{r}{right hand list}
}
\description{
This function encodes a season variable based on dates of data collection in the study.
inspired by http://stackoverflow.com/questions/7519790/assign-multiple-new-variables-on-lhs-in-a-single-line-in-r.
}
