% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs}
\alias{gs}
\title{sp polygons with green spaces}
\format{The @data contains the following variables:
\describe{
  \item{id}{Unique polygon ID}
  \item{type}{Type of green space}
}}
\usage{
gs
}
\description{
A sptialPolygonsDataFrame object with green space in Doncaster
}
\keyword{datasets}
