% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repel.R
\name{haplotype_labels}
\alias{haplotype_labels}
\title{Group Haplotype Taxon Labels}
\usage{
haplotype_labels(haplotypes, id_column)
}
\arguments{
\item{haplotypes}{Joined haplotypes with labels read in by \code{\link[=join_bayes]{join_bayes()}}}

\item{id_column}{Character vector for indirect tidy evaluation.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} of multi-taxa node labels with \code{x}, \code{y}, \code{label},
and \code{fill} \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} aesthetics.
}
\description{
Create labels of multiple taxa tip positions grouped by count.
}
\examples{
list.files(
  path = system.file("extdata/MrBayes", package = "thesis"),
  pattern = "rps-infile.nex.con.tre",
  full.names = TRUE
) \%>\%
  thesis::read_tree(tree_file = .) \%>\%
  thesis::join_bayes(
    tree_data = .,
    id_column = "Taxon_a_posteriori",
    scale_vector = c(5, 10)
  ) \%>\%
  haplotype_labels(haplotypes = ., id_column = "Taxon_a_posteriori")

}
\keyword{internal}
