% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hawaii}
\alias{hawaii}
\title{sp object to plot Hawaiian archipelago}
\format{The @data contains the following variables:
\describe{
  \item{ID_1}{State ID}
  \item{NAME_1}{State name}
}}
\source{
\url{http://biogeo.ucdavis.edu/data/diva/adm/USA_adm.zip}
}
\usage{
hawaii
}
\description{
A spatialPolygonsDataFrame object to plot the Hawaiian archipelago.
}
\keyword{datasets}
