% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-plot.R
\name{heat.map.plot}
\alias{heat.map.plot}
\title{Creates a separation plot for diagnosing logistic regression models.}
\usage{
heat.map.plot(data, col.actual, col.probs)
}
\description{
\code{separation.plot} Create a separation plot based on paper.
}

