% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_wrappers.R
\name{html_caption}
\alias{html_caption}
\title{Replace LaTeX Italics}
\usage{
html_caption(chunk_type, caption)
}
\arguments{
\item{chunk_type}{Knitr chunk type returned by
\code{opts_knit$get("rmarkdown.pandoc.to")}}

\item{caption}{Character scalar passed to \code{knitr::kable()} \code{caption} arg.}
}
\value{
Caption formatted by knitr pandoc conversion type.
}
\description{
Convert LaTeX italics to HTML markdown for firgure captions.
}
\keyword{internal}
