% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{italicize}
\alias{italicize}
\title{LaTeX Italicization}
\usage{
italicize(unitalicized, chunk_type)
}
\arguments{
\item{unitalicized}{Character vector to italicize.}
}
\description{
Wrapper to italicize vector based on \code{knitr} chunk type.
}
\examples{
italicize(unitalicized = "Physaria floribunda", chunk_type = "latex")

}
\seealso{
Other text: 
\code{\link{capitalize}()}
}
\concept{text}
\keyword{internal}
