% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{jitter_violin}
\alias{jitter_violin}
\title{Built violin plots with jittered points}
\usage{
jitter_violin(
  specimen_tbl,
  trait,
  aesthetic_id,
  aesthetic_labels = NULL,
  legend_title = NULL,
  violin.params = list(),
  jitter.params = list(),
  theme.params = list()
)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{trait}{Variable with trait observation in
\link{herbarium_specimens}.}

\item{aesthetic_id}{Variable to set color and shape aesthetics of jitters.}

\item{aesthetic_labels}{Named vector of labels for manual scale values.}

\item{legend_title}{Character vector of length one to set the ggplot
legend title.}

\item{violin.params}{List of arguments passed to [ggplot2::geom_violin().}

\item{jitter.params}{List of arguments passed to \code{\link[ggplot2:geom_jitter]{ggplot2::geom_jitter()}}.}

\item{theme.params}{List of arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
\link{ggplot} of trait observations delineated by reviewed id (x-axis)
and jittered points (y-axis) representing numeric values.
}
\description{
Semi-flexible plotting of continuous trait violin plots with adjustable
point jitters.
}
\keyword{internal}
