% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\name{join_bayes}
\alias{join_bayes}
\title{Join MrBayes Import with Labels and Geoms}
\usage{
join_bayes(tree_data, id_column, ...)
}
\arguments{
\item{tree_data}{Tibble output by \code{\link[=read_tree]{read_tree()}} with ggtree node and
label variables.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}

\item{...}{Forwarding arguments to \code{\link[=node_geoms]{node_geoms()}} (e.g., \code{scale_vector}).}
}
\value{
Tibble of multi-sample split node labels and geom sizes joined to
MrBayes import by \code{\link[=read_tree]{read_tree()}} and \link{dna_specimens} data.
}
\description{
Combine condensed node labels for multi-sample tips from \code{\link[=node_labels]{node_labels()}}
with scaled geom sizes from \code{\link[=node_geoms]{node_geoms()}}.
}
\examples{
joined_ggtree <-
  list.files(system.file("extdata/MrBayes", package = "thesis"),
    full.names = TRUE, pattern = "rITS-infile.nex.con.tre"
  ) \%>\%
  read_tree(tree_file = .) \%>\%
  join_bayes(tree_data = ., id_column = "prior_id")

}
\keyword{internal}
