% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kauai}
\alias{kauai}
\title{sp object to plot Kauai}
\format{The @data contains the following variables:
\itemize{
  \item{ID_1}{State ID}
  \item{NAME_1}{State name}
  \item{ID_2}{County ID}
  \item{NAME_2}{County name}
}}
\source{
Subset from \url{http://biogeo.ucdavis.edu/data/diva/adm/USA_adm.zip}
}
\usage{
kauai
}
\description{
A spatialPolygonsDataFrame object of Kauai, Hawaii.
}
\keyword{datasets}
