% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_wrappers.R
\name{knitr_section}
\alias{knitr_section}
\title{Knitr Section Writer}
\usage{
knitr_section(knitr_title, knitr_type, chunk_type)
}
\arguments{
\item{knitr_title}{Character vector of length one matching the section title.}

\item{knitr_type}{Character vector of length one matching the section type.}

\item{chunk_type}{Knitr chunk type returned by
\code{opts_knit$get("rmarkdown.pandoc.to")}}
}
\value{
Text to stdout with section formatting based on pandoc conversion.
}
\description{
Function to return text to stdout based on the knit option of markdown to
pandoc conversion for either "html" or "latex" output.  For composing
Bookdown documents in HTML or LaTeX output.
}
\keyword{internal}
