% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knitr_table_dna}
\alias{knitr_table_dna}
\title{Table 2 Bookdown}
\usage{
knitr_table_dna(chunk_type)
}
\arguments{
\item{chunk_type}{Knitr chunk type returned by
\code{opts_knit$get("rmarkdown.pandoc.to")}}
}
\value{
Table 2 reference based on type of pandoc conversion.
}
\description{
Methods table reference for DNA specimens.
}
\keyword{internal}
