% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lads}
\alias{lads}
\title{sp object of local authority districts (LADs) in England and Wales}
\format{The @data contains the following variables:
\describe{
  \item{code}{Local authority district (LAD) code}
  \item{name}{LAD name}
  \item{region}{Government office region}
  \item{supergroup}{Supergroup area classification of the LAD}
  \item{group}{Group area classification of the LAD}
  \item{subgroup}{Subgroup classification of the LAD}
  \item{lexp_female}{Life expectancy, females}
  \item{lexp_male}{Life expectancy, males}
  \item{lim_little}{Day-to-day activities limited a little}
  \item{lim_lot}{Day-to-day activities limited a lot}
  \item{lim_not}{Day-to-day activities not limited}
  \item{v_bad}{Self-reported general health very bad}
  \item{bad}{Self-reported general health bad}
  \item{fair}{Self-reported general health fair}
  \item{good}{Self-reported general health good}
  \item{v_good}{Self-reported general health very good}
  \item{imd}{Index of multiple deprivation (average for LAD)}
  \item{imd_rank}{Rank of index of multiple deprivation (average for LAD)}
  \item{unem}{Number of economically active unemployed people at census day 2011}
  \item{pop_1000}{Population of LADs in 1,000s (based on srgh)}
  \item{bad_over}{Self-reported general health bad overall (bad + v_bad)}
  \item{good_over}{SRGH good overall (good + v_good)}
}}
\source{
\itemize{
  \item Boundary data: \url{https://census.edina.ac.uk/easy_download_data.html?data=infuse_dist_lyr_2011}
  \item Area classification: \url{http://www.ons.gov.uk/ons/guide-method/geography/products/area-classifications/ns-area-classifications/ns-2011-area-classifications/datasets/2011-census-data.xls}
  \item Life expectancy: ONS; Public Health England
  \item Limiting long-term illness or disability: Census (Nomis)
  \item Self-reported general health: Census (Nomis)
  \item IMD: Dept. for Communities and Local Government
  \item Economically active unemployed from the census 2011 (Nomis)
  \item Population: calculated from census (2011)
}
}
\usage{
lads
}
\description{
Local authority districts (LADs) in England and Wales saved as a polygons
spatial object, including area classifications.
}
\keyword{datasets}
