% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-maps.R
\name{layer_borders}
\alias{layer_borders}
\title{Add state and county borders \code{LayerInstance}}
\usage{
layer_borders(
  spl_extent,
  sf_crs = NULL,
  sf_state_color = "black",
  sf_county_color = "white",
  ...
)
}
\arguments{
\item{spl_extent}{A data frame with specimen layer extent returned by
\link{spl_bbox}}

\item{sf_crs}{Optional CRS string called by \link[sf:st_transform]{st_transform}}

\item{sf_state_color}{Color of simple features boundry layer.}

\item{sf_county_color}{Color of simple features boundry layer.}

\item{...}{Forwarded arguments from \link{build_map}}
}
\value{
A ggplot2 `LayerInstance`` for border boundary plotting.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Build ggplot layer instance for state and county borders.
Boundary shapefiles are downloaded from the US Census Bureau
Topologically Integrated Geographic Encoding and Referencing (TIGER)
database and read into R with the \link{tigris} package (Walker 2016).
Data are read in to R as simple features through the \link{sf} package
(Pebesma 2018) and transformed with the \link{rmapshaper} package
(Teucher and Russell 2020) to simplify border polygons using the
Visvalingam algorithm.
}
\examples{
# tigris::tigris_cache_dir(path = 'PATH TO MY NEW CACHE DIRECTORY')
# Sys.getenv('TIGRIS_CACHE_DIR')
# options(tigris_class = "sf", tigris_use_cache = TRUE)
ggplot2::ggplot() +
  layer_borders(spl_extent = spl_bbox(thesis::spp_co_front_range))

}
\references{
Pebesma E. 2018. Simple feature for R: Standardized support for spatial
vector data. The R Journal. 10(1):438-446.

Teucher A, Russell K. 2020. rmapshaper: Client for 'mapshaper' for
'Geospatial' operations. R package version 0.4.4.
https://CRAN.R-project.org/package=rmapshaper

Walker K. 2016. tigris: An R package to access and work with
geographic data from the US census bureau. The R Journal. 8(2):231-242.
}
\keyword{internal}
