% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-maps.R
\name{layer_elevation}
\alias{layer_elevation}
\title{Layer elevation raster for specimen plot.}
\usage{
layer_elevation(specimen_tbl, raster_zoom = 7, raster_factor = 2, ...)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{raster_zoom}{Integer vector of length 1 for zoom level passed to
\code{elevatr::get_elev_raster()} function call.  Must be between 1 and 14.}

\item{raster_factor}{Integer vector of length 1 for optional aggregation
of elevation raster to decrease resolution / object size.}

\item{...}{Forwarded arguments from \link{build_map}}
}
\value{
Base ggplot layer from \link{elevatr} raster object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Build base ggplot layer from digital elevation model data from AWS Open Data
Terrain Tiles via the \link{elevatr} package.
Package repository: https://github.com/jhollist/elevatr
Map terrain tiles: https://registry.opendata.aws/terrain-tiles/
}
\examples{
library(ggplot2)
elev_ggplot <-
  layer_elevation(
    specimen_tbl = thesis::spp_co_front_range,
    raster_zoom = 7, raster_factor = 1
  )
elev_ggplot +
  layer_specimens(
    specimen_tbl = spp_co_front_range,
    id_column = "Taxon_a_posteriori"
  ) +
  coord_sf(
    xlim = range(spl_bbox(elev_ggplot)$Longitude),
    ylim = range(spl_bbox(elev_ggplot)$Latitude)
  ) +
  labs(x = "Longitude", y = "Latitude")

}
\references{
Hollister, J.W., Tarak Shah (2017). elevatr: Access Elevation Data from
Various APIs.
}
\keyword{internal}
