% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-maps.R
\name{layer_specimens}
\alias{layer_specimens}
\title{Add specimen \code{LayerInstance}}
\usage{
layer_specimens(
  specimen_tbl,
  id_column,
  legend_status = TRUE,
  shape_aes = FALSE,
  geom_size = 3,
  jitter_width = NULL,
  jitter_height = NULL,
  jitter_alpha = 1
)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}

\item{legend_status}{Optional logical vector to show legend via \code{show.legend}
argument of \link[ggplot2:geom_point]{geom_point}.}

\item{shape_aes}{Logical to optionally add shape aesthetic to ggplot.}

\item{geom_size}{Numeric scalar for size of jitter aesthetic.}

\item{jitter_width}{Numeric scalar for \code{\link[=geom_jitter]{geom_jitter()}} width.}

\item{jitter_height}{Numeric scalar for \code{\link[=geom_jitter]{geom_jitter()}} height.}

\item{jitter_alpha}{Numeric scalar for \code{\link[=geom_jitter]{geom_jitter()}} alpha value.}
}
\value{
A ggplot2 `LayerInstance`` for specimen plotting.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Build a ggplot2 layer instance with aesthetics determined by \code{id_column}
matching a variable in \code{specimen_tbl}. Geom point positions are determined
by longitude and latitude coordinates, with color and optional shape
aesthetics determined by the values in \code{id_column}.
}
\examples{
ggplot2::ggplot() +
  layer_specimens(
    specimen_tbl = thesis::spp_co_front_range,
    id_column = "Taxon_a_posteriori", shape_aes = FALSE,
    jitter_width = 0.033, jitter_height = 0.033
  )

}
\keyword{internal}
