% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-maps.R
\name{layer_themes}
\alias{layer_themes}
\title{Add map ggplot scales and themes}
\usage{
layer_themes(specimen_tbl, id_column, legend_title)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}

\item{legend_title}{Character vector of length one to set the ggplot
legend title.}
}
\value{
List of scale layers (colour, shape), themes, and axis labels.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Add manual values to specimen identification discrete scales in map ggplots
set by \link{spp_color} and \link{spp_shape} vectors. Modify plot themes
and add markdown italics for specific epithets with \link{spl_labels}.
}
\examples{
# Add theme specifications and markdown legend.
ggplot2::ggplot() +
  layer_borders(
    spl_extent = spl_bbox(spp_co_front_range),
    sf_county_color = "black"
  ) +
  layer_specimens(
    specimen_tbl = spp_co_front_range, shape_aes = TRUE,
    id_column = "Taxon_a_posteriori"
  ) +
  layer_themes(
    specimen_tbl = spp_co_front_range,
    id_column = "Taxon_a_posteriori",
    legend_title = "Reviewed Annotations"
  )

}
\keyword{internal}
