% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-nearest.R
\name{left_join.by.nearest}
\alias{left_join.by.nearest}
\title{Left join on dataframes by mapping nearest point in data to each query point.}
\usage{
\method{left_join}{by.nearest}(query.df, data.df, query.x = "long",
  query.y = "lat", data.x = "long", data.y = "lat", dist = FALSE)
}
\description{
\code{left_join.by.nearest} returns nearest point from data coordinates for each query
coordinates.
}

