% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leisure_centres}
\alias{leisure_centres}
\title{sp object containing locations of Doncaster leisure centres in BNG}
\format{The @data slot contains the following variables:
\describe{
  \item{name}{Leisure centre name}
}}
\source{
Google Places API
}
\usage{
leisure_centres
}
\description{
sp spatialPointsDataFrame object of leisure centres in Doncaster in
British National Grid projection
}
\keyword{datasets}
