% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{map_trait_distribution}
\alias{map_trait_distribution}
\title{Trait Distribution Map}
\usage{
map_trait_distribution(
  tidy_trait,
  bb_xlim = c(-114, -102.5),
  bb_ylim = c(37.5, 48.75)
)
}
\arguments{
\item{tidy_trait}{Tibble of specimen trait distributions.}

\item{bb_xlim}{Numeric vctor to scale \code{sf} coordinates x-limits.}

\item{bb_ylim}{Numeric vctor to scale \code{sf} coordinates y-limits.}
}
\value{
A \code{ggplot} with specimen trait distribution mapping.
}
\description{
Given a trait subset (e.g., discrete or continuous morphological characters),
build a \code{ggplot} over state and county border layers returned by
\code{\link[=layer_borders]{layer_borders()}} and with standardized theme and coordinates.
}
\examples{
thesis::herbarium_specimens \%>\%
  dplyr::filter(!is.na(.data$Ovule_number)) \%>\%
  dplyr::bind_cols(
    .,
    range_split(trait_tbl = ., split_var = "Ovule_number")
  ) \%>\%
  dplyr::rename(Trait = "Ovule_number_max") \%>\%
  dplyr::mutate(Trait = as.factor(x = .data$Trait)) \%>\%
  map_trait_distribution(tidy_trait = .) +
  ggplot2::scale_color_viridis_d(option = "D") +
  ggplot2::labs(color = "Max Ovule Count per Locule")

}
\keyword{internal}
