% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-nearest.R
\name{match.nearest}
\alias{match.nearest}
\title{Get nearest data point for each query point.}
\usage{
match.nearest(query.x, query.y, data.x, data.y)
}
\description{
\code{sum} returns nearest point from data coordinates for each query
coordinates.
}

