% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation_minvol_weight.R
\docType{data}
\name{minvol_weight}
\alias{minvol_weight}
\title{Industry weight of MSCI Min Volatility indices from 10/31/2011 to 01/05/2017}
\format{A data frame with 11 variables 
\itemize{ 
  \item Sector = Sector of asset
  \item Date = Date when asset was in the portfolio
  \item Weight = Weight of the sector (as of the released date)
  }}
\description{
Industry weight of MSCI Min Volatility indices from 10/31/2011 to 01/05/2017
}
\keyword{asset}
\keyword{data}
\keyword{min}
\keyword{msci}
\keyword{security}
\keyword{volatility}

