% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_wrappers.R
\name{msa_texshade}
\alias{msa_texshade}
\title{Format TexShade Environment}
\usage{
msa_texshade(tex_path, landscape = FALSE)
}
\arguments{
\item{tex_path}{MSA subset \emph{.tex} file path}

\item{landscape}{Logical scalar to add LaTeX landscape page formatting.
Used for alignment intervals after first 165 residues to accomodate
chapter headings.}
}
\description{
Output \emph{.tex} files generated by the R package msa (Bodenhofer et al. 2015)
define a TeXshade (Beitz 2000) environment for displaying multiple sequence
alignments. Additional formatting to alignment text size and optional
LaTeX landscape page orientation.
}
\details{
Given an \emph{.tex} file output by \code{\link[msa:msaPrettyPrint]{msa::msaPrettyPrint()}}, the
TEXshade environment text is subset and the contents overwritten to include
sequence alignment character sizing and formatting.
}
\keyword{internal}
