% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_wrappers.R
\name{msa_wrapper}
\alias{msa_wrapper}
\title{Subset \code{\link[msa:msaPrettyPrint]{msa::msaPrettyPrint()}} \emph{.tex} Alignments}
\usage{
msa_wrapper(
  msa_fasta,
  msa_output,
  alignment_dir = fs::path("Appendix/alignments/")
)
}
\arguments{
\item{msa_fasta}{File name of \verb{extdata/Alignments/} system file.
FASTA alignments available as installed external data files.}

\item{msa_output}{Locus name to identify output \emph{.tex} files.}

\item{alignment_dir}{Directory path to write \emph{.tex} MSA subset files.}
}
\description{
From installed external data system files, read in alignment from FASTA
file, calculate alignment intervals, and write subset \emph{.tex} output files.
}
\details{
\itemize{
\item Default subdirectory to write \emph{.tex} alignments: \verb{Appendix/alignments/}
}
}
\keyword{internal}
