% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{munge}
\alias{munge}
\title{Munge}
\usage{
munge(df)
}
\arguments{
\item{df}{matrix or data table}
}
\value{
lowercase data frame with white space trimmed. numeric, Date, and integer columns are left alone.
}
\description{
Quick and dirty simple munging for data frames with mostly character data.
}
\details{
many thanks to this stack overflow answer: http://stackoverflow.com/questions/17964513/subset-variables-in-data-frame-based-on-column-type
}
\examples{
df = data.frame(number = numeric(c(1,2,3)), characternumber = c("one ", " TwO", "ThrEe"), city = c(" Memphis", "MEMPHIS", "memphis "))
munge(df)
}
\keyword{data}
\keyword{munging}

