% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\name{node_geoms}
\alias{node_geoms}
\title{Scale Node Geoms}
\usage{
node_geoms(tree_data, id_column, scale_vector = c(4, 12))
}
\arguments{
\item{tree_data}{Tibble output by \code{\link[=read_tree]{read_tree()}} with ggtree node and
label variables.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}

\item{scale_vector}{Numeric vector of length two passed to
\code{\link[scales:rescale]{scales::rescale()}} \code{to} argument.}
}
\value{
Tibble with \code{node}, \code{id_column}, and \code{geom_size} variables
where \code{geom_size} represents geom sizes scaled by \code{node} and \code{id_column}.
}
\description{
Group \code{\link[ggtree:ggtree]{ggtree::ggtree()}} samples by \code{node} and \code{id_column} to
scale geom point sizes for tree tips.
}
\examples{
list.files(system.file("extdata/MrBayes", package = "thesis"),
  full.names = TRUE, pattern = "rITS-infile.nex.con.tre"
) \%>\%
  read_tree(tree_file = .) \%>\%
  node_geoms(tree_data = ., id_column = "prior_id")

}
\keyword{internal}
