% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\name{node_labels}
\alias{node_labels}
\title{Label Multi-taxa Nodes}
\usage{
node_labels(tree_data)
}
\arguments{
\item{tree_data}{Tibble output by \code{\link[=read_tree]{read_tree()}} with ggtree node and
label variables.}
}
\value{
Tibble with \code{node}, \code{single_label}, \code{label}, and \code{node_group}
variables where \code{node_group} represents multi-sample terminal nodes.
}
\description{
For tips (i.e. terminal nodes) with multiple taxa split into
\code{single_label} by \link{read_tree}, create a new variable
\code{node_group} to distinguish multi-sample genotype labels.
}
\examples{
list.files(system.file("extdata/MrBayes", package = "thesis"),
  full.names = TRUE, pattern = "rITS-infile.nex.con.tre"
) \%>\%
  read_tree(tree_file = .) \%>\%
  node_labels(tree_data = .)

}
\keyword{internal}
