% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nunique.R
\name{nunique}
\alias{nunique}
\title{nunique}
\usage{
nunique(x)
}
\arguments{
\item{x}{vector or column of a dataframe}
}
\value{
atomic number listing number of unique values in x
}
\description{
Get number of unique values in a vector or column of a dataframe
}
\examples{
vec = c(1,3,4,2,5,5,5,1,1,3,7,9)
nunique(vec)
}
