% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oa_comm_2001_lookup}
\alias{oa_comm_2001_lookup}
\title{Data frame containing lookups between OAs and COMM areas for 2001}
\format{The data frame containts the following variables:
\describe{
  \item{oacode}{Output area code}
  \item{codecomm}{Community area code}
  \item{namecomm}{Community area name}
}}
\usage{
oa_comm_2001_lookup
}
\description{
Data frame containing lookups between output areas and community areas
for Doncaster for 2001 geographies
}
\keyword{datasets}
