% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-themes.R
\name{parse_taxa}
\alias{parse_taxa}
\title{Parse Taxa Labels}
\usage{
parse_taxa(tree_tibble, id_column)
}
\arguments{
\item{tree_tibble}{\code{tbl_tree} object with label vector \code{id_column}.}

\item{id_column}{Character vector for indirect tidy evaluation.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Pair with \code{\link[=haplotype_labels]{haplotype_labels()}} and \code{\link[=species_plot]{species_plot()}} to create plotmath
expressions parsed as \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} labels.

Plotmath Reference
\itemize{
\item https://rdrr.io/r/grDevices/plotmath.html
}
}
\examples{
system.file("extdata/BEAST/spp-hypothesis-1.mcc", package = "thesis") \%>\%
  treeio::read.beast(file = .) \%>\%
  ggtree::fortify() \%>\%
  dplyr::mutate(
    label = gsub("medicinae", "'medicinae'", x = .data$label)
  ) \%>\%
  dplyr::mutate(
    label = parse_taxa(tree_tibble = ., id_column = "label")
  ) \%>\%
  dplyr::select(label)

}
\seealso{
Other labels: 
\code{\link{repel_haplotype_labels}()},
\code{\link{repel_map_labels}()},
\code{\link{repel_node_labels}()}
}
\concept{labels}
\keyword{internal}
