% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pilot_ind}
\alias{pilot_ind}
\title{Pilot individual level (survey) data}
\format{The data frame contains the following variables:
\describe{
  \item{pidp}{Unique person ID}
  \item{sex}{Sex at most recent wave}
  \item{qual}{Highest qualification at most recent wave}
  \item{eth}{Ethnic group}
  \item{ten}{Tenure at most recent wave}
  \item{car}{Number of cars at most recent wave}
  \item{age}{Age at most recent wave}
  \item{llid}{Limiting long--term illness or disability at most recent wave}
}}
\source{
\url{https://discover.ukdataservice.ac.uk/catalogue/?sn=6614\&type=Data\%20catalogue}
}
\usage{
pilot_ind
}
\description{
Individual level survey data from Understanding Society for pilot simulation.
}
\keyword{datasets}
