% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pilot_msoa_ext}
\alias{pilot_msoa_ext}
\title{Pilot simulation extracted weights of health conditions at MSOA level.}
\format{The data frame contains the following variables:
\describe{
  \item{code}{Doncaster LSOA code}
  \item{total}{Simulated total population}
  \item{sex_female}{count by sex}
  \item{sex_male}{count by sex}
  \item{qual_0}{count by highest qualification}
  \item{qual_2}{count by highest qualification}
  \item{qual_3}{count by highest qualification}
  \item{qual_4_plus}{count by highest qualification}
  \item{qual_other}{count by highest qualification}
  \item{eth_british}{count by ethnicity}
  \item{eth_irish}{count by ethnicity}
  \item{eth_other_white}{count by ethnicity}
  \item{eth_mixed_multiple_ethnic}{count by ethnicity}
  \item{eth_asian_asian_british}{count by ethnicity}
  \item{eth_black_african_caribbean_british}{count by ethnicity}
  \item{eth_other_ethnicity}{count by ethnicity}
  \item{ten_owned_outright}{count by tenure}
  \item{ten_owned_mortgage_shared}{count by tenure}
  \item{ten_rented}{count by tenure}
  \item{car_0}{count by car ownership}
  \item{car_1}{count by car ownership}
  \item{car_2_plus}{count by car ownership}
  \item{age_16_17}{count by age}
  \item{age_18_19}{count by age}
  \item{age_20_24}{count by age}
  \item{age_25_29}{count by age}
  \item{age_30_44}{count by age}
  \item{age_45_59}{count by age}
  \item{age_60_64}{count by age}
  \item{age_65_74}{count by age}
  \item{age_75_84}{count by age}
  \item{age_85_89}{count by age}
  \item{age_90_plus}{count by age}
  \item{llid_no}{simulated count with LLID}
  \item{llid_yes}{simulated count with LLID}
}}
\source{
2011 Census Tables from Nomisweb; Understanding Society.
}
\usage{
pilot_msoa_ext
}
\description{
Data frame of results of pilot spatial microsimulation for health conditions
at middle layer super ouput area (MSOA) level.
}
\keyword{datasets}
