% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pilot_oa_int}
\alias{pilot_oa_int}
\title{Pilot simulation integerised results of health conditions at output area
level.}
\format{The data frame contains the following variables:
\describe{
  \item{pidp}{ID from Understanding Society}
  \item{sex}{Sex (male or female)}
  \item{qual}{Highest qualification}
  \item{eth}{Ethnicity (reduced categories)}
  \item{ten}{Housing tenure}
  \item{car}{Car ownership}
  \item{age}{Age (banded categories)}
  \item{llid}{Limiting long--term illness or disability}
  \item{zone}{Output area}
}}
\source{
2011 Census Tables from Nomisweb; Understanding Society.
}
\usage{
pilot_oa_int
}
\description{
Data frame of results of pilot spatial microsimulation for health conditions
at ouput area level.
}
\keyword{datasets}
