% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_wrappers.R
\name{potentially_informative_sites}
\alias{potentially_informative_sites}
\title{Potentially Informative Site Alignment}
\usage{
potentially_informative_sites(fasta_file, ...)
}
\arguments{
\item{fasta_file}{FASTA file to read \link{DNAMultipleAlignment} from.}

\item{...}{Forwarding arguments to \code{\link[ggmsa:ggmsa]{ggmsa::ggmsa()}}}
}
\value{
\code{ggplot} object with multiple sequence alignment.
}
\description{
Create multiple sequence alignment ggplot with \code{\link[ggmsa:ggmsa]{ggmsa::ggmsa()}} from multiple
sequence alignment subset to potentially informative sites.
FASTA files are read into R via \code{\link[Biostrings:MultipleAlignment-class]{Biostrings::readDNAMultipleAlignment()}}
}
\examples{
\dontrun{
ml_fasta <-
  list.files("data-raw/sequencing/3.alignments-subset",
    full.names = TRUE
  )
potentially_informative_sites(fasta_file = ml_fasta[[1]])
}

}
\keyword{internal}
