% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob-var-plot.R
\name{prob_var_plot}
\alias{prob_var_plot}
\title{LOESS Plot for Binary Variable}
\usage{
prob_var_plot(data, x_var, y_var, window_width, method = "bayes")
}
\arguments{
\item{data}{The dataframe containing the binary and continuous variable}

\item{x_var}{The name of the continuous variable (as a string)}

\item{y_var}{The name of the binary response variable (as a string)}

\item{window_width}{Width of the window to use for approximating p.}

\item{method}{The method to be used to find confidence intervals for
binomial variables. See documentation for binom.confint for options.}
}
\value{
A ggplot2 plot.
}
\description{
\code{prob_var_plot} creates a ggplot2 plot with a Bernoulli LOESS for
binary versus continuous variable. This should be useful as a way to
determine the relationship between a continuous variable and a binary
variable.
}
\details{
It returns a ggplot2 plot, so one could easily add other elements to the
plot, such as titles, axes labels, and other themes.
}
\examples{
x <- rnorm(100, 0, 2)
y <- rbinom(100, 1, invlogit(x))
test_data <- data.frame(x, y)
p <- prob_var_plot(test_data, "x", "y", 2)
p
p + labs(title="Empirical probability versus Gaussian variable")
}

