% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{range_split}
\alias{range_split}
\title{Split Numeric Ranges}
\usage{
range_split(trait_tbl, split_var)
}
\arguments{
\item{trait_tbl}{Tibble of specimen observations to split range data from.}

\item{split_var}{Single column tibble of range data in character vector.}
}
\value{
Tibble (3 columns) with character vector of raw ranges and
two numeric vectors of min / max values from ranges split by "-".
}
\description{
Given a tibble of range data in a character vector, split ranges separated
by "-" character.  Creates two new tibble columns containing numeric vectors
with minimum and maximum sorted numeric values split from character range.
}
\examples{
herbarium_specimens \%>\%
  dplyr::select(Ovule_number) \%>\%
  dplyr::filter(!is.na(Ovule_number)) \%>\%
  range_split(trait_tbl = ., split_var = "Ovule_number")
}
\keyword{internal}
