% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_weight.R
\name{rel_weight}
\alias{rel_weight}
\title{This function calculates the relative weight of each sector for each day 
and display the result in a dataframe}
\usage{
rel_weight(category)
}
\description{
This function calculates the relative weight of each sector for each day 
and display the result in a dataframe
}

