% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repel.R
\name{repel_haplotype_labels}
\alias{repel_haplotype_labels}
\title{Repel Tree Labels}
\usage{
repel_haplotype_labels(
  tree_nudges,
  tree_labels,
  initial_ggtree,
  label_size = 2
)
}
\arguments{
\item{tree_nudges}{Inline \code{\link[tibble:tribble]{tibble::tribble()}} with variables to set
\code{nudge_x}, \code{nudge_y}, \code{segment.curvature}, \code{color} (label text),
and variables to join by:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  - `node`: Numeric scalar denoting phylogenetic tree node.
  - `Taxon_a_posteriori`: Character scalar of reviewed identification.
}\if{html}{\out{</div>}}}

\item{tree_labels}{Tibble of labels output by \code{\link[=haplotype_labels]{haplotype_labels()}} with
with \link{ggplot} aesthetics for \code{x}, \code{y}, \code{label}, and \code{fill} aesthetics.}

\item{initial_ggtree}{Base layer \code{\link[ggtree:ggtree]{ggtree::ggtree()}} for \code{\link[purrr:map]{purrr::map()}} and
\code{\link[purrr:reduce]{purrr::reduce()}} into \code{\link[rlang:call2]{rlang::call2()}} to build R call expressions.}

\item{label_size}{Numeric scalar to indicate \code{size} parameter for
the repelled geom layer.}
}
\value{
List of R call expressions to layer repelled labels with coordinate
nudges onto \code{\link[ggtree:ggtree]{ggtree::ggtree()}} object.
}
\description{
Leverage \code{\link[rlang:call2]{rlang::call2()}} to build call expressions, mapping
a \code{\link[tibble:tribble]{tibble::tribble()}} with x- and y-axis nudges for layered
labels via \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel()}}. For expression evaluation,
pipe output into \code{\link[rlang:eval_tidy]{rlang::eval_tidy()}}.
}
\details{
For nodes with multiple taxa in the base layer \code{\link[ggtree:ggtree]{ggtree::ggtree()}}, repelled
labels denote specimen annotations with counts tabulated by
\code{\link[=haplotype_labels]{haplotype_labels()}}.
}
\examples{
bayes_joined <-
  list.files(
    path = system.file("extdata/MrBayes", package = "thesis"),
    pattern = "rps-infile.nex.con.tre",
    full.names = TRUE
  ) \%>\%
  thesis::read_tree(tree_file = .) \%>\%
  thesis::join_bayes(
    tree_data = .,
    id_column = "Taxon_a_posteriori",
    scale_vector = c(5, 10)
  )

bayes_haplotypes <-
  thesis::haplotype_labels(
    haplotypes = bayes_joined,
    id_column = "Taxon_a_posteriori"
  )

bayes_ggtree <-
  ggtree::ggtree(tr = bayes_joined, layout = "circular") +
  ggplot2::geom_point(mapping = ggplot2::aes(color = prior_id), size = 3)

tibble::tribble(
  ~"nudge_x", ~"nudge_y", ~"segment.curvature", ~"node", ~"Taxon_a_posteriori", ~"color",
  0.01, -1.5, 0.1, 5, "Physaria 'medicinae'", "white",
  0.0075, -1.5, 0.1, 5, "Physaria didymocarpa subsp. didymocarpa", "white",
  0.005, -1.5, 0.1, 5, "Physaria eburniflora", "white",
  0.0025, -1.5, 0.1, 5, "Physaria acutifolia", "black"
) \%>\%
  thesis::repel_haplotype_labels(
    tree_nudges = .,
    tree_labels = bayes_haplotypes,
    initial_ggtree = bayes_ggtree,
    label_size = 3
  ) \%>\%
  rlang::eval_tidy(expr = .)

}
\seealso{
haplotype_labels

Other labels: 
\code{\link{parse_taxa}()},
\code{\link{repel_map_labels}()},
\code{\link{repel_node_labels}()}
}
\concept{labels}
\keyword{internal}
