% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repel.R
\name{repel_node_labels}
\alias{repel_node_labels}
\title{Repel Node Probabilities}
\usage{
repel_node_labels(node_nudges, node_labels, initial_ggtree, label_size = 3)
}
\arguments{
\item{node_nudges}{Inline \code{\link[tibble:tribble]{tibble::tribble()}} with variables to set
\code{nudge_x}, \code{nudge_y}, \code{segment.curvature}
and variable to join by:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  - `node`: Numeric scalar denoting phylogenetic tree node.
}\if{html}{\out{</div>}}}

\item{node_labels}{Tibble subset from labels read in by \code{\link[=join_bayes]{join_bayes()}} to
nodes with posterior probabilities.}

\item{initial_ggtree}{Base layer \code{\link[ggtree:ggtree]{ggtree::ggtree()}} for \code{\link[purrr:map]{purrr::map()}} and
\code{\link[purrr:reduce]{purrr::reduce()}} into \code{\link[rlang:call2]{rlang::call2()}} to build R call expressions.}

\item{label_size}{Numeric scalar to indicate \code{size} parameter for
the repelled geom layer.}
}
\value{
List with \code{\link[ggplot2:ggplot2-package]{ggplot2::ggplot2()}} \code{LayerInstance} for node repelled
posterior probabilities by tree node.
}
\description{
Leverage \code{\link[rlang:call2]{rlang::call2()}} to build call expressions, mapping
a \code{\link[tibble:tribble]{tibble::tribble()}} with x- and y-axis nudges for layered
labels via \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel()}}. For expression evaluation,
pipe output into \code{\link[rlang:eval_tidy]{rlang::eval_tidy()}}.
}
\details{
\code{\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel()}} \code{LayerInstance} for repelled node posterior
probabilities. Excludes nodes with missing values. Probabilites are rounded
to 3 digits.
}
\examples{
haplotypes <-
  list.files(
    path = system.file("extdata/MrBayes", package = "thesis"),
    pattern = "rps-infile.nex.con.tre",
    full.names = TRUE
  ) \%>\%
  thesis::read_tree(tree_file = .) \%>\%
  thesis::join_bayes(
    tree_data = .,
    id_column = "Taxon_a_posteriori"
  )

probabilities <- haplotypes \%>\%
  dplyr::filter(.data$prob != 1) # Exclude MrBayes polytomies

bayes_ggtree <-
  ggtree::ggtree(tr = haplotypes, layout = "circular") +
  ggplot2::geom_point(mapping = ggplot2::aes(color = prior_id), size = 3)

tibble::tribble(
  ~"nudge_x", ~"nudge_y", ~"segment.curvature", ~"node",
  -0.001, 0, 0.01, 31,
  -0.002, 0, 0.01, 32,
  -0.003, 0, 0.01, 33,
  -0.004, 0, 0.01, 34,
  -0.005, 0, 0.01, 35
) \%>\%
  thesis::repel_node_labels(
    node_nudges = .,
    node_labels = probabilities,
    initial_ggtree = bayes_ggtree,
    label_size = 4
  ) \%>\%
  rlang::eval_tidy(expr = .)

}
\seealso{
Other labels: 
\code{\link{parse_taxa}()},
\code{\link{repel_haplotype_labels}()},
\code{\link{repel_map_labels}()}
}
\concept{labels}
\keyword{internal}
