% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_con}
\alias{res_con}
\title{Data frame containing constraints for microsimulation at OA level}
\format{The data frame contains the following variables:
\describe{
  \item{code}{Output area code}
  \item{sex_female}{count by sex}
  \item{sex_male}{count by sex}
  \item{qual_0}{count by highest qualification}
  \item{qual_2}{count by highest qualification}
  \item{qual_3}{count by highest qualification}
  \item{qual_4_plus}{count by highest qualification}
  \item{qual_other}{count by highest qualification}
  \item{eth_british}{count by ethnicity}
  \item{eth_irish}{count by ethnicity}
  \item{eth_other_white}{count by ethnicity}
  \item{eth_mixed_multiple_ethnic}{count by ethnicity}
  \item{eth_asian_asian_british}{count by ethnicity}
  \item{eth_black_african_caribbean_british}{count by ethnicity}
  \item{eth_other_ethnicity}{count by ethnicity}
  \item{ten_owned_outright}{count by tenure}
  \item{ten_owned_mortgage_shared}{count by tenure}
  \item{ten_rented}{count by tenure}
  \item{car_0}{count by car ownership}
  \item{car_1}{count by car ownership}
  \item{car_2_plus}{count by car ownership}
  \item{age_16_17}{count by age}
  \item{age_18_19}{count by age}
  \item{age_20_24}{count by age}
  \item{age_25_29}{count by age}
  \item{age_30_44}{count by age}
  \item{age_45_59}{count by age}
  \item{age_60_64}{count by age}
  \item{age_65_74}{count by age}
  \item{age_75_84}{count by age}
  \item{age_85_89}{count by age}
  \item{age_90_plus}{count by age}
  \item{eca_emp}{Economic activity: Employed (full-time or part-time)}
  \item{eca_homefam}{Economic activity: looking after home or family}
  \item{eca_ltsick}{Economic activity: long-term sick}
  \item{eca_other}{Economic activity: other}
  \item{eca_retired}{Economic activity: retired}
  \item{eca_selfemp}{Economic activity: self-employed}
  \item{eca_student}{Economic activity: student}
  \item{eca_unemp}{Economic activity: unemployed}
  \item{mar_civil_part}{Civil partnership}
  \item{mar_divorced}{Divorced}
  \item{mar_married}{Married}
  \item{mar_separated}{Separated (still legally married/civil partnership)}
  \item{mar_single}{Single (never married)}
  \item{mar_widowed}{Widowed}
}}
\source{
2011 Census Tables from Nomisweb
}
\usage{
res_con
}
\description{
Data frame containing constraints for microsimulation.
Tables are for output areas in Doncaster (978).
First column contain the zone (OA) codes, all subsequent columns contain
counts.
}
\keyword{datasets}
