% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_ind}
\alias{res_ind}
\title{Data frame containing individual-level data for resilience simulation.}
\format{The data frame contains the following variables:
\describe{
  \item{pidp}{Unique person ID}
  \item{car}{Number of cars at most recent wave}
  \item{ten}{Tenure at most recent wave}
  \item{qual}{Highest qualification at most recent wave}
  \item{mar}{Marital status}
  \item{eca}{Economic activity}
  \item{sex}{Sex at most recent wave}
  \item{eth}{Ethnic group}
  \item{age}{Age at most recent wave}
  \item{depress}{Clinical depression self-reported}
  \item{nhood_coh}{Neighbourhood cohesion (similar to others in neighbourhood)}
  \item{nhood_trust}{Neighbourhood trust}
  \item{nhood_belong}{Belong to neighbourhood}
  \item{fin_sit}{Subjective current financial situation}
  \item{ghq_conc}{GHQ: concentration}
  \item{ghq_sleep}{GHQ: sleep loss}
  \item{ghq_useful}{GHQ: felt useful}
  \item{ghq_decision}{GHQ: can make decisions}
  \item{ghq_strain}{GHQ: Under strain}
  \item{ghq_diff}{GHQ: Problems overcoming difficulties}
  \item{ghq_enjoy}{GHQ: enjoy day-to-day activities}
  \item{ghq_face}{GHQ: able to face problems}
  \item{ghq_unhappy}{GHQ: unhappy or depressed}
  \item{ghq_confid}{GHQ: losing confidence}
  \item{ghq_worth}{GHQ: believe worthless}
  \item{ghq_happy}{GHQ: general happiness}
  \item{kids}{Number of children aged 0-15 in household}
  \item{hh_income}{Gross household income month before interview}
  \item{poverty}{Is household in poverty (less than 60\% median income)?}
  \item{alcohol}{Alcohol consumption}
}}
\source{
\url{https://discover.ukdataservice.ac.uk/catalogue/?sn=6614\&type=Data\%20catalogue}
}
\usage{
res_ind
}
\description{
Data frame containing individual-level survery data for resilience
simulation.
}
\keyword{datasets}
