% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_weights_ext}
\alias{res_weights_ext}
\title{Simulation (fractional weight) results of health resilience at output
area level.}
\format{The data frame contains the following variables:
\describe{
  \item{code}{Doncaster output area code}
  \item{total}{Simulated total population}
  \item{car_0}{count by car ownership}
  \item{car_1}{count by car ownership}
  \item{car_2_plus}{count by car ownership}
  \item{ten_owned_outright}{count by tenure}
  \item{ten_owned_mortgage_shared}{count by tenure}
  \item{ten_rented}{count by tenure}
  \item{qual_0}{count by highest qualification}
  \item{qual_2}{count by highest qualification}
  \item{qual_3}{count by highest qualification}
  \item{qual_4_plus}{count by highest qualification}
  \item{qual_other}{count by highest qualification}
  \item{mar_civil_part}{Marital status: civil partnership}
  \item{mar_divorced}{Marital status: divorced}
  \item{mar_married}{Marital status: married}
  \item{mar_separated}{Marital status: separated}
  \item{mar_single}{Marital status: single (never married)}
  \item{mar_widowed}{Marital status: Widowed or widower}
  \item{eca_emp}{Employed}
  \item{eca_homefam}{Looking after home or family}
  \item{eca_ltsick}{Long-term sick or disabled}
  \item{eca_other}{Other economic activity}
  \item{eca_retired}{Retired}
  \item{eca_selfemp}{Self-employed}
  \item{eca_student}{Student}
  \item{eca_unemp}{Unemployed (looking for work or ready to start job)}
  \item{sex_female}{count by sex}
  \item{sex_male}{count by sex}
  \item{eth_asian_asian_british}{count by ethnicity}
  \item{eth_black_african_caribbean_british}{count by ethnicity}
  \item{eth_british}{count by ethnicity}
  \item{eth_irish}{count by ethnicity}
  \item{eth_mixed_multiple_ethnic}{count by ethnicity}
  \item{eth_other_ethnicity}{count by ethnicity}
  \item{eth_other_white}{count by ethnicity}
  \item{age_16_17}{count by age}
  \item{age_18_19}{count by age}
  \item{age_20_24}{count by age}
  \item{age_25_29}{count by age}
  \item{age_30_44}{count by age}
  \item{age_45_59}{count by age}
  \item{age_60_64}{count by age}
  \item{age_65_74}{count by age}
  \item{age_75_84}{count by age}
  \item{age_85_89}{count by age}
  \item{age_90_plus}{count by age}
  \item{depress_no}{No depression}
  \item{depress_yes}{Clinical depression}
  \item{ncoh_no}{Neighbourhood cohesion: no}
  \item{ncoh_yes}{Neighbourhood cohesion}
  \item{trust_no}{Neighbourhood trust: no}
  \item{trust_yes}{Neighbourhood trust}
  \item{belong_no}{Do not belong to neighbourhood}
  \item{belong_yes}{Belong to neighbourhood}
  \item{fin_bad}{Subjective financial situation: not good}
  \item{fin_good}{Subjective financial situation: managing or comfortable}
  \item{ghq_conc_bad}{GHQ concentration: bad}
  \item{ghq_conc_good}{GHQ concentration: good}
  \item{ghq_sleep_bad}{GHQ sleep loss: bad}
  \item{ghq_sleep_good}{GHQ sleep loss: good}
  \item{ghq_useful_bad}{GHQ play useful role: bad}
  \item{ghq_useful_good}{GHQ play useful role: good}
  \item{ghq_decision_bad}{GHQ can make decisions: bad}
  \item{ghq_decision_good}{GHQ can make decisions: good}
  \item{ghq_strain_bad}{GHQ constantly under strain: bad}
  \item{ghq_strain_good}{GHQ constantly under strain: good}
  \item{ghq_diff_bad}{GHQ problems overcoming difficulties: bad}
  \item{ghq_diff_good}{GHQ problems overcoming difficulties: good}
  \item{ghq_enjoy_bad}{GHQ enjoy day-to-day activities: bad}
  \item{ghq_enjoy_good}{GHQ enjoy day-to-day activities: good}
  \item{ghq_face_bad}{GHQ able to face problems: bad}
  \item{ghq_face_good}{GHQ able to face problems: good}
  \item{ghq_unhappy_bad}{GHQ unhappy or depressed: bad}
  \item{ghq_unhappy_good}{GHQ unhappy or depressed: good}
  \item{ghq_confid_bad}{GHQ losing confidence: bad}
  \item{ghq_confid_good}{GHQ losing confidence: good}
  \item{ghq_worth_bad}{GHQ believe worthless: bad}
  \item{ghq_worth_good}{GHQ believe worthless: good}
  \item{ghq_happy_bad}{GHQ general happiness: bad}
  \item{ghq_happy_good}{GHQ general happiness: good}
  \item{alcohol_high}{Alcohol consumption: high risk(>14 units/week)}
  \item{alcohol_low}{Alcohol consumption: low risk (<=14 units/week)}
}}
\source{
Spatial microsimulation from 2011 Census Tables from Nomisweb and
Understanding Society.
}
\usage{
res_weights_ext
}
\description{
Data frame of fractional weights from resilience spatial microsimulation for
health conditions at ouput area level.
}
\keyword{datasets}
