% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plot}
\alias{save_plot}
\title{Save Figures}
\usage{
save_plot(gg_plot, ...)
}
\arguments{
\item{gg_plot}{Input plot to save ggplot graphic \emph{.png} and \emph{.pdf} figures.}

\item{...}{Forwarding arguments to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\description{
Compiled PDF and HTML documents require different graphics devices.
To facilitate \code{Makefile} automation, save both \emph{.pdf} and \emph{.png} files
from input ggplot. Non-standard evaluation with \code{rlang} is used for naming
files from the input \code{ggplot} object name.
}
\keyword{internal}
