% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seinet_coords}
\alias{seinet_coords}
\title{SEINet Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 564 rows and 20 columns.
}
\source{
\url{https://swbiodiversity.org/seinet/index.php}

\describe{
\item{scientificName}{Identification by scientific name}
\item{scientificNameAuthorship}{Taxon author}
\item{genus}{Genus}
\item{specificEpithet}{Specific epithet}
\item{taxonRank}{Taxon rank (e.g. var., subsp.)}
\item{infraspecificEpithet}{Infraspecific epithet}
\item{identifiedBy}{Record identifier}
\item{recordedBy}{Record collector}
\item{recordNumber}{Collection number}
\item{eventDate}{Date of collection}
\item{year}{Year of collection}
\item{stateProvince}{State}
\item{county}{County}
\item{locality}{Description of collection locality}
\item{Latitude}{Degree decimal latitude}
\item{Longitude}{Degree decimal longitude}
\item{verbatimCoordinates}{Verbatim coordinates: TRS, WGS}
\item{verbatimElevation}{}
\item{recordId}{SEINet unique record identifier}
\item{references}{SEINet record site reference}
}
}
\usage{
seinet_coords
}
\description{
SEINet records for \emph{P. floribunda} and \emph{P. bellii}.
}
\keyword{datasets}
