% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{separate_discrete_trait}
\alias{separate_discrete_trait}
\title{Separate Discrete Trait Observations}
\usage{
separate_discrete_trait(specimen_tbl, trait_selection, filter_n = 20)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{trait_selection}{Tibble variable to split discrete observations from.}

\item{filter_n}{Cutoff for filtering rarely observed traits by count.}
}
\value{
Tibble of specimen observations separated into rows by unique trait
observations.
}
\description{
Separate records of discrete trait observations from a variable where
multiple values may be observed in a single voucher.
}
\examples{
separate_discrete_trait(
  specimen_tbl = herbarium_specimens,
  trait_selection = "Replum_shape"
)

}
\keyword{internal}
