% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation-plot.R
\name{separation.plot}
\alias{separation.plot}
\title{Creates a separation plot for diagnosing logistic regression models.}
\usage{
separation.plot(data, col.actual, col.probs)
}
\description{
\code{separation.plot} Create a separation plot based on paper.
}

