% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimens.R
\name{Specimen}
\alias{Specimen}
\title{Voucher Specimens}
\description{
R6 class defined as a specimen record data structure. The \link{Specimen}
superclass holds a \code{\link[tibble:tibble]{tibble}} object containing voucher
metadata. Public methods allow for reporting the number of distinct and
duplicate records, subsetting specimen records by geographic coordinate,
and filtering operations by either annotation or collector / collection.
}
\examples{

## ------------------------------------------------
## Method `Specimen$new`
## ------------------------------------------------

# Construct object instance with `$new()` method
voucher <- Specimen$new(
  records = thesis::herbarium_specimens,
  identifier = "Taxon_a_posteriori"
)

# Basic access to the public `records` field tibble
class(voucher$records)
dim(voucher$records)

## ------------------------------------------------
## Method `Specimen$census`
## ------------------------------------------------

# Check basic record census of total and distinct specimens.
voucher$census()

## ------------------------------------------------
## Method `Specimen$limit`
## ------------------------------------------------

# Subset records by geographic coordinates
limits <- voucher$clone()
limits$limit(west = -107, east = -105, south = 39, north = 41)

dim(limits$records)

## ------------------------------------------------
## Method `Specimen$taxa`
## ------------------------------------------------

# Filter specimens by taxonomic rank; here, reviewed specific epithets.
species <- voucher$clone()
species$taxa(c("flori", "medi"), identifier = "Taxon_a_posteriori")

# Index tibble records column annotations to tabulate.
table(species$records[["Taxon_a_posteriori"]])

## ------------------------------------------------
## Method `Specimen$collections`
## ------------------------------------------------

found <- voucher$clone()
found$collections(5068, "Rollins" = c(5145, 5146), "Mulligan")
found$census()

dplyr::select(found$records, "Collector", "Collection_Number")

## ------------------------------------------------
## Method `Specimen$annotations`
## ------------------------------------------------

limits$annotations()

## ------------------------------------------------
## Method `Specimen$labels`
## ------------------------------------------------

dna_vouchers <- dplyr::select(thesis::dna_specimens, ID_final) \%>\%
 dplyr::rename(label = ID_final) \%>\%
 Specimen$new(
   records = .,
   identifier = "ID_final"
 )

# Note length of the returned label vector equals the record row number.
length(dna_vouchers$labels())
unique(dna_vouchers$labels())
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{records}}{A \code{\link[tibble:tbl_df-class]{tbl_df}} S3 tibble object
with set of specimen vouchers.}

\item{\code{identifier}}{Annotation variable in a \code{\link[=Specimen]{Specimen$records}}
tibble. Used to designate \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} scale labels.}

\item{\code{seed}}{Random seed value for deterministic reproducibility.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Specimen-new}{\code{Specimen$new()}}
\item \href{#method-Specimen-census}{\code{Specimen$census()}}
\item \href{#method-Specimen-limit}{\code{Specimen$limit()}}
\item \href{#method-Specimen-taxa}{\code{Specimen$taxa()}}
\item \href{#method-Specimen-collections}{\code{Specimen$collections()}}
\item \href{#method-Specimen-annotations}{\code{Specimen$annotations()}}
\item \href{#method-Specimen-labels}{\code{Specimen$labels()}}
\item \href{#method-Specimen-clone}{\code{Specimen$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-new"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-new}{}}}
\subsection{Method \code{new()}}{
Construct a \code{Specimen} container
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$new(
  records,
  identifier = NULL,
  seed = format(Sys.Date(), format = "\%Y\%m\%d")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{records}}{Specimen voucher records \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{\code{identifier}}{Character scalar for voucher annotation label variable.}

\item{\code{seed}}{Random seed for reproducibility.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Construct object instance with `$new()` method
voucher <- Specimen$new(
  records = thesis::herbarium_specimens,
  identifier = "Taxon_a_posteriori"
)

# Basic access to the public `records` field tibble
class(voucher$records)
dim(voucher$records)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-census"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-census}{}}}
\subsection{Method \code{census()}}{
Record census accounting of voucher specimens.
Reports two summations:
\itemize{
\item Total number of specimens \strong{including duplicate collections}
\item Number of \emph{distinct} specimen collections (i.e., excluding duplicates)
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$census()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Check basic record census of total and distinct specimens.
voucher$census()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-limit"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-limit}{}}}
\subsection{Method \code{limit()}}{
Limit records by geographic coordinate (lon/lat).
This method enables filtering records by minimum or maximum
coordinate limits using any combination of the four cardinal directions.
Each parameter sets the directional bound (min/max) to filter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$limit(west = NULL, east = NULL, south = NULL, north = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{west}}{Filter records by \strong{minimum longitude} (min. x)}

\item{\code{east}}{Filter records by \strong{maximum longitude} (max. x)}

\item{\code{south}}{Filter records by \strong{minimum latitude} (min. y)}

\item{\code{north}}{Filter records by \strong{maximum latitude} (max. y)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Subset records by geographic coordinates
limits <- voucher$clone()
limits$limit(west = -107, east = -105, south = 39, north = 41)

dim(limits$records)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-taxa"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-taxa}{}}}
\subsection{Method \code{taxa()}}{
Given a \href{#public-fields}{\code{Specimen$records}} tibble, utilize
\emph{partial} string matching for taxonomic filtering of specimen records.
Flexibility is provided to query for e.g., multiple specific epithets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$taxa(..., identifier = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Taxonomic terms to filter
records against column designation by \code{identifier} parameter.}

\item{\code{identifier}}{\code{records} field tibble variable for regular expression
matching by \code{\link[dplyr:filter]{dplyr::filter()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Filter specimens by taxonomic rank; here, reviewed specific epithets.
species <- voucher$clone()
species$taxa(c("flori", "medi"), identifier = "Taxon_a_posteriori")

# Index tibble records column annotations to tabulate.
table(species$records[["Taxon_a_posteriori"]])
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-collections"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-collections}{}}}
\subsection{Method \code{collections()}}{
Query specimen records by collector or collection number.
This method uses dot collection via \code{\link[rlang:list2]{rlang::dots_list()}}, allowing for
flexible specification of various collector and collection combos.
Input arguments should take one of the following forms:
\itemize{
\item Numeric vector of collection numbers named by collector
\item Character scalar specifying a single collector string
\item Numeric scalar specifying a single collection number
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$collections(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Any number of collector
collections (i.e., numeric vector named by collector), collectors
(character scalars) or collection numbers (numeric scalars).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[tibble:tbl_df-class]{tbl_df}} subset of matched
collector / collection specimen records.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{found <- voucher$clone()
found$collections(5068, "Rollins" = c(5145, 5146), "Mulligan")
found$census()

dplyr::select(found$records, "Collector", "Collection_Number")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-annotations"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-annotations}{}}}
\subsection{Method \code{annotations()}}{
Create markdown-formatted specimen annotations.
Used for plotting \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} manual scale values with italicized
fonts via the \code{\link[ggtext:element_markdown]{ggtext::element_markdown()}} extension.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$annotations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named list of markdown- (i.e., HTML) formatted annotations.
Length equals number of unique values in the \code{identifier} field.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{limits$annotations()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-labels"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-labels}{}}}
\subsection{Method \code{labels()}}{
Create base expression specimen labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$labels(.identifier = "label")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.identifier}}{Character scalar denoting \code{records} field variable
for filtering and annotation operations involving taxonomic identifications.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character vector of expressions for parsed font labels.
Length equivalent to length of the records \code{label} variable.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{dna_vouchers <- dplyr::select(thesis::dna_specimens, ID_final) \%>\%
 dplyr::rename(label = ID_final) \%>\%
 Specimen$new(
   records = .,
   identifier = "ID_final"
 )

# Note length of the returned label vector equals the record row number.
length(dna_vouchers$labels())
unique(dna_vouchers$labels())
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
