% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-maps.R
\name{spl_bbox}
\alias{spl_bbox}
\title{Specimen layer bounding box}
\usage{
spl_bbox(...)
}
\arguments{
\item{...}{One of specimen tibble or ggplot base layer.}
}
\value{
Data frame of Longitude and Latitude bounding box coordinates.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Calculate bounding box from specimen tibbles or ggplot base layers.
}
\examples{
spl_bbox(herbarium_specimens)

ggmap_ggplot <- layer_ggmap(
  specimen_tbl = thesis::spp_co_front_range,
  gg_map_type = "satellite"
)
spl_bbox(ggmap_ggplot)
spl_bbox(spp_co_front_range)

}
\keyword{internal}
