% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-maps.R
\name{spl_id}
\alias{spl_id}
\title{Map Individual Specimen}
\usage{
spl_id(
  specimen_tbl,
  id_column,
  collector,
  collection,
  shape_aes = NULL,
  geom_size = 3,
  label_size = 3,
  h_adjust = 0.25,
  v_adjust = -0.15
)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}

\item{collector}{Character scalar to filter specimen by collector name.}

\item{collection}{Numeric vector to filter specimens by collection number.}

\item{shape_aes}{Logical to optionally add shape aesthetic to ggplot.}

\item{geom_size}{Numeric scalar for size of jitter aesthetic.}

\item{label_size}{Numeric vector of length one for label size.}

\item{h_adjust}{Numeric vector of length one for horizontal label adjustment.}

\item{v_adjust}{Numeric vector of length one for vertical label adjustment.}
}
\value{
ggplot object with added specimen annotation layer.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Generate geom specifications to add label of individual specimen(s).
}
\examples{
ggplot2::ggplot() +
  layer_specimens(
    specimen_tbl = spp_co_front_range,
    id_column = "prior_id", shape_aes = "prior_id"
  ) +
  spl_id(
    specimen_tbl = spp_co_front_range,
    id_column = "prior_id", shape_aes = "prior_id",
    collector = "Wolf", collection = 642
  )

}
\keyword{internal}
