% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-themes.R
\name{spl_labels}
\alias{spl_labels}
\title{Generate HTML vector for
\link[ggtext:element_markdown]{element_markdown()}}
\usage{
spl_labels(specimen_tbl, id_column)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}
}
\value{
Character vector of html markup named by specimen identification.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Requires installation of \link{ggtext} as follows:
remotes::install_github("clauswilke/ggtext")
}
\examples{
spl_labels(
  specimen_tbl = spp_co_front_range,
  id_column = "prior_id"
)

}
\keyword{internal}
