% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-maps.R
\name{spl_legend}
\alias{spl_legend}
\title{Extract specimen layer legend}
\usage{
spl_legend(specimen_tbl, id_column, legend_title, shape_aes, geom_size)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}

\item{legend_title}{Character vector of length one to set the ggplot
legend title.}

\item{shape_aes}{Logical to optionally add shape aesthetic to ggplot.}

\item{geom_size}{Numeric scalar for size of jitter aesthetic.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Build a minimal specimen layer ggplot with themes and extract the legend.
}
\examples{
cowplot::plot_grid(
  spl_legend(
    specimen_tbl = spp_co_front_range,
    id_column = "Taxon_a_posteriori",
    legend_title = "Review Annotations",
    shape_aes = TRUE, geom_size = 3
  )
)

}
\keyword{internal}
