% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-maps.R
\name{spl_order}
\alias{spl_order}
\title{Arrange Grouped Specimens}
\usage{
spl_order(specimen_tbl, id_column)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{id_column}{Character scalar matching specimen tibble ID variable.
Variable used for aesthetic (color and shape) mappings.}
}
\value{
Tibble data frame sorted by grouped specimen identifications.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Order specimens for ggplot layers grouped by identification column. ggplot
layers are determined by row number, where groups with higher counts are
arranged first.
}
\examples{
library(dplyr)
distinct(spp_co_front_range, .data$prior_id)
spp_co_front_range \%>\%
  group_by(prior_id) \%>\%
  tally(x = .)
spl_order(specimen_tbl = spp_co_front_range, id_column = "prior_id") \%>\%
  distinct(.data$prior_id)

}
\keyword{internal}
