% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-maps.R
\name{spl_states}
\alias{spl_states}
\title{Bounding box State intersect}
\usage{
spl_states(spl_extent, sf_states)
}
\arguments{
\item{spl_extent}{A data frame with specimen layer extent returned by
\link{spl_bbox}}

\item{sf_states}{State simple features data frame returned by
\link[tigris:states]{states}}
}
\value{
Character vector of states intersecting with spatial extent layer.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Build a bounding box polygon simple feature from a spatial extent data frame
returned by \link{spl_bbox}. Given a set of state simple features, identify
geometry intersections with \link[sf:geos_binary_pred]{st_intersects} to subset state simple
features for plotting.
}
\examples{
states_sfs <- tigris::states() \%>\% rmapshaper::ms_simplify(input = .)
spl_bbox(herbarium_specimens) \%>\%
  spl_states(spl_extent = ., sf_states = states_sfs)

}
\keyword{internal}
