% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spp_co_front_range}
\alias{spp_co_front_range}
\title{CO Front Range Subset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 97 rows and 66 columns.
}
\usage{
spp_co_front_range
}
\description{
Used for Roxygen @examples and testthat unit test. \cr
Bounding box:
\itemize{
\item Longitude: c(-107.5, -104.5)
\item Latitude: c(38, 41)
}
}
\keyword{internal}
