% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spp_color}
\alias{spp_color}
\title{Species Color Aesthetics}
\format{
An object of class \code{character} of length 24.
}
\usage{
spp_color
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\describe{
\item{spp_color}{Vector with values for ggplot color aesthetics}
}
}
\keyword{internal}
