% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spp_shape}
\alias{spp_shape}
\title{Species Shape Aesthetics}
\format{
An object of class \code{numeric} of length 24.
}
\usage{
spp_shape
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\describe{
\item{spp_shape}{Vector with values for ggplot shape aesthetics}
}
}
\keyword{internal}
