% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-specimens.R
\name{subset_coords}
\alias{subset_coords}
\title{Subset specimen data by geographic coordinates.}
\usage{
subset_coords(specimen_tbl, Latitude = NULL, Longitude = NULL)
}
\arguments{
\item{specimen_tbl}{Data.frame with columns Latitude / Longitude.
Should be numeric vectors of decimal degree coordinates.}

\item{Latitude}{Character vector length two min / max latitude.}

\item{Longitude}{Character vector length two min / max longitude.}
}
\value{
Data.frame of specimens within coordinate bounds.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Given an input data.frame and at least one of latitude or longitude
vectors with minimum and maximum values to subset coordinate data.
}
\examples{
specimen_subset <- subset_coords(
  specimen_tbl = herbarium_specimens,
  Latitude = c(41, 45),
  Longitude = c(-109, -105)
)
range(specimen_subset$Latitude)
range(specimen_subset$Longitude)

}
\keyword{internal}
