% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sy_county}
\alias{sy_county}
\title{sp object to plot South Yorkshire}
\format{An object of class \code{SpatialPolygons} of length 1.}
\source{
\url{https://census.edina.ac.uk/easy_download_data.html?data=England_ct_2011}
}
\usage{
sy_county
}
\description{
A spatialPolygons object (i.e. no DataFrame) to plot the boundary of the
historic county of South Yorkshire to show Doncaster in context
}
\keyword{datasets}
