% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis-package.R
\docType{package}
\name{thesis-package}
\alias{thesis}
\alias{thesis-package}
\title{thesis: Data and functions for systematic botany thesis work}
\description{
Data and code for compiling a thesis manuscript. Includes functions for importing and parsing herbarium and DNA specimen information, distribution mapping, phylogenetic tree annotation, and quantitative trait comparisons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jasonratcliff/thesis}
  \item Report bugs at \url{https://github.com/jasonratcliff/thesis/issues}
}

}
