% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyembase.R
\name{tidyembase}
\alias{tidyembase}
\title{tidyembase}
\usage{
tidyembase(filepath)
}
\arguments{
\item{filepath}{path to an exported list of citations from embase. filepath must include search strings.}
}
\value{
cleaned data frame
}
\description{
This function encodes a season variable based on dates of data collection in the study.
}
\examples{
tidyembase('path/to/file.csv')
tidyembase('path/to/file.xlsx') 
}
\keyword{csv,}
\keyword{embase}
\keyword{excel,}
