% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us}
\alias{us}
\title{Understanding Society data set}
\format{The data frame contains the following variables:
\describe{
  \item{pidp}{Unique person ID}
  \item{f_hidp}{f_ household id}
  \item{e_hidp}{e_ household id}
  \item{d_hidp}{d_ household id}
  \item{c_hidp}{c_ household id}
  \item{b_hidp}{b_ household id}
  \item{a_hidp}{a_ household id}
  \item{region}{Government Office Region at latest wave}
  \item{age_full}{Age (individual years)}
  \item{age}{Age at most recent wave, cut into bands to match census tables}
  \item{sex}{Sex at most recent wave}
  \item{eth_full}{Ethnic group}
  \item{eth_minority}{Is respondent White British or minority (TRUE)}
  \item{eth}{Ethnicity, reduced categories to match census}
  \item{qual}{Highest qualification at most recent wave}
  \item{car}{Number of cars at most recent wave}
  \item{heat}{Central heating at most recent wave}
  \item{ten_full}{Tenure at most recent wave}
  \item{ten}{Housing tenure at most recent wave, reduced categories to match census}
  \item{econ_act}{Economic activity}
  \item{ocrowd}{Live in overcrowded house (more than one person per room)?}
  \item{llid}{Limiting long--term illness or disability at most recent wave}
  \item{srgh}{Self--reported general health at most recent wave}
  \item{class8}{Class (NS--SEC 8) at most recent wave}
  \item{marital}{Marital status}
  \item{couple}{Cohabiting as part of a couple (married or otherwise)?}
  \item{nhood_coh}{Neighbourhood cohesion (similar to others in neighbourhood)}
  \item{nhood_trust}{Neighbourhood trust}
  \item{nhood_belong}{Belong to neighbourhood}
  \item{civic}{Civic participation (active in community org)}
  \item{social_coh}{Social cohesion (people get along)}
  \item{safe_alone}{Feel safe going out alone at night}
  \item{sim_inc}{Proportion friends similar income}
  \item{sim_eth}{Proportion friends similar ethnicity}
  \item{pol_eff}{Political efficacy: make a difference}
  \item{fin_sit}{Subjective current financial situation}
  \item{soc_isol}{Socially isolated}
  \item{ghq_conc}{GHQ: concentration}
  \item{ghq_sleep_loss}{GHQ: sleep loss}
  \item{ghq_useful}{GHQ: felt useful}
  \item{ghq_decision}{GHQ: can make decisions}
  \item{ghq_strain}{GHQ: Under strain}
  \item{ghq_difficult}{GHQ: Problems overcoming difficulties}
  \item{ghq_enjoy}{GHQ: enjoy day-to-day activities}
  \item{ghq_face}{GHQ: able to face problems}
  \item{ghq_unhappy}{GHQ: unhappy or depressed}
  \item{ghq_confidence}{GHQ: losing confidence}
  \item{ghq_selfworth}{GHQ: believe worthless}
  \item{ghq_happiness}{GHQ: general happiness}
  \item{cig}{Smoke cigarettes: yes or no}
  \item{alcohol}{Alcohol consumption}
  \item{bf_dist}{Distance to closest friend (<1 mile)}
  \item{bf_often}{Frequency see best friend {>1/week}}
  \item{save}{Save money?}
  \item{nkids}{Number of children aged <16 in the household (integer)}
  \item{kids}{Number of children aged <16 in the household (factor)}
  \item{hh_income}{Gross household income month before interview (top-coded
  to £20,000 per month)}
  \item{cohab}{Number of cohabitants (i.e. hhsize - respondent)}
  \item{depress}{Clinical depression}
  \item{ihd}{Ischemic heart disease or stroke (air pollution marker)}
  \item{prob_pay}{Problems paying housing costs (rent/mortgage)}
  \item{hh_cost}{Monthly housing cost (rent/mortgage)}
  \item{uc_replace}{Receives benefit being replaced by Universal Credit
  (JSA, income support, ESA, tax credits)?}
}}
\source{
\url{https://discover.ukdataservice.ac.uk/catalogue/?sn=6614\&type=Data\%20catalogue}
}
\usage{
us
}
\description{
Understanding Society data frame
}
\keyword{datasets}
