% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation_usa.R
\docType{data}
\name{usa}
\alias{usa}
\title{MSCI USA data from 10/31/2011 to 01/05/2017}
\format{A data frame with 11 variables 
\itemize{ 
  \item Ticker = The symbol of each security.
  \item Name = The name of each security.
  \item Asset Class = The asset class of each security.
  \item Weight = The weight of each asset relative to the entire portfolio
  \item Price = Spot price of asset 
  \item Shares = Number of shares of asset in portfolio
  \item Market Value = Market Value of asset at spot price
  \item Notional Value = Total amount of security's underlying asset value at spot price
  \item Sector = Sector of asset
  \item SEDOL = Stock Exchange Daily Official List Number
  \item ISIN = International Securities Identification Number
  \item Exchange = Stock Exchange where asset is traded
  \item Date = Date when asset was in the portfolio
  }}
\description{
This data frame contains the MSCI USA indices and their consitutents from 
2011-10-31 to 2017-05-05.
}
\keyword{data}
\keyword{min}
\keyword{msci}
\keyword{usa}

