% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yh_region}
\alias{yh_region}
\title{sp object outline of the Yorkshire and The Humber region}
\format{The @data contains the following variables:
\describe{
  \item{code}{Geography code}
  \item{name}{Name or label of the geography}
  \item{label}{Geography code}
}}
\source{
\url{https://census.edina.ac.uk/easy_download_data.html?data=England_gor_2011}
}
\usage{
yh_region
}
\description{
A spatialPolygonsDataFrame object of the outline of the Yorkshire and The
Humber region, subset from all English government office regions (GOR).
}
\keyword{datasets}
