% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_weight.R
\name{compute_rel_weight}
\alias{compute_rel_weight}
\title{This function calculates the relative weight of each sector for each day 
and display the result in a dataframe
export}
\usage{
compute_rel_weight()
}
\description{
This function calculates the relative weight of each sector for each day 
and display the result in a dataframe
export
}

