% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repel.R
\name{repel_map_labels}
\alias{repel_map_labels}
\title{Repel Map Labels}
\usage{
repel_map_labels(map_nudges, map_labels, initial_ggplot)
}
\arguments{
\item{map_nudges}{Inline \code{\link[tibble:tribble]{tibble::tribble()}} with variables to set
\code{nudge_x}, \code{nudge_y}, \code{segment.curvature}, and variable to join by:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  - `Key`: e.g., collector(s) and collection separated by "_".
}\if{html}{\out{</div>}}}

\item{map_labels}{A \code{\link[tibble:tibble]{tibble::tibble()}} input for \code{ggplot} with \code{x}, \code{y}, and
\code{label} aesthetics passed to \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel()}}.}

\item{initial_ggplot}{Base layer \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} for \code{\link[purrr:map]{purrr::map()}} and
\code{\link[purrr:reduce]{purrr::reduce()}} into \code{\link[rlang:call2]{rlang::call2()}} to build R call expressions.}
}
\value{
List of R call expressions to layer repelled labels with coordinate
nudges onto \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Leverage \code{\link[rlang:call2]{rlang::call2()}} to build call expressions, mapping
a \code{\link[tibble:tribble]{tibble::tribble()}} with x- and y-axis nudges for layered
labels via \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel()}}. For expression evaluation,
pipe output into \code{\link[rlang:eval_tidy]{rlang::eval_tidy()}}.
}
\details{
For each call layered onto a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}, labels represent
specimen sampling identifying individual records of interest.
}
\examples{
extent <-
  tibble::tribble(
    ~"Longitude", ~"Latitude",
    -108, 39.75,
    -105, 39.75,
    -108, 41.75,
    -105, 41.75
  )

specimens <- thesis::herbarium_specimens \%>\%
  subset_coords(
    specimen_tbl = .,
    Longitude = c(-108, -105),
    Latitude = c(39.75, 41.75)
  )

jackson <-
  thesis::find_spp(
    specimen_tbl = thesis::herbarium_specimens,
    collector = "Kastning|Nelson",
    collection = "1462|1725|49286|49478"
  ) \%>\%
  dplyr::select(Collector, Collection_Number, Longitude, Latitude) \%>\%
  dplyr::mutate(
    Label = stringr::str_remove_all(
      string = Collector,
      pattern = "[A-Z]\\\\. ?"
    ) \%>\%
      gsub("with|and", "&", x = .) \%>\%
      paste(., Collection_Number, sep = "\n"),
    Collection_Number = as.numeric(.data$Collection_Number),
    Key = stringr::str_replace_all(
      string = .data$Label,
      pattern = "[^A-z0-9]+",
      replacement = "_"
    )
  )

map <- ggplot2::ggplot() +
  thesis::layer_specimens(
    specimen_tbl = specimens,
    id_column = "Taxon_a_posteriori"
  ) +
  ggplot2::geom_point(
    data = jackson,
    mapping = ggplot2::aes(x = Longitude, y = Latitude),
    shape = 5, color = "black", size = 4
  ) +
  ggplot2::coord_sf(
    xlim = range(extent$Longitude),
    ylim = range(extent$Latitude)
  ) +
  ggplot2::theme_classic()

tibble::tribble(
  ~"nudge_x", ~"nudge_y", ~"segment.curvature", ~"Key",
  0.4, -0.05, 0.1, "Nelson_49286",
  -1, -0.25, 0.1, "Kastning_Kastning_1462",
  -1, -0.25, 0.1, "Kastning_Culp_1725",
  -0.4, -0.1, -0.1, "Nelson_49478",
) \%>\%
  thesis::repel_map_labels(
    map_nudges = .,
    map_labels = jackson,
    initial_ggplot = map
  ) \%>\%
  rlang::eval_tidy(expr = .)

}
\seealso{
Other labels: 
\code{\link{parse_taxa}()},
\code{\link{repel_haplotype_labels}()},
\code{\link{repel_node_labels}()}
}
\concept{labels}
\keyword{internal}
