% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{trait_phenology}
\alias{trait_phenology}
\title{Plot Trait Phenology}
\usage{
trait_phenology(
  specimen_tbl,
  trait,
  aesthetic_id,
  aesthetic_labels = NULL,
  legend_title = NULL,
  jitter.params = list(),
  theme.params = list()
)
}
\arguments{
\item{specimen_tbl}{Tibble of herbarium specimens subset.}

\item{trait}{Variable with trait observation in
\link{herbarium_specimens}.}

\item{aesthetic_id}{Variable to set color and shape aesthetics of jitters.}

\item{aesthetic_labels}{Named vector of labels for manual scale values.}

\item{legend_title}{Character vector of length one to set the ggplot
legend title.}

\item{jitter.params}{List of arguments passed to \code{\link[ggplot2:geom_jitter]{ggplot2::geom_jitter()}}.}

\item{theme.params}{List of arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
\link{ggplot} object of trait phenology by "\%m-\%d" date format.
}
\description{
Wrapper to \code{\link[=ggplot]{ggplot()}} layering points by collection date for a given
y variable speecified by \code{trait}. The \code{aesthetic_id} determines
\strong{color} and \strong{shape} geom point aesthetics and \code{aesthetic_labels}
are passed as manual scale specifications. Use \code{legend_title} to set
the combined color and shape scale legend title.
}
\details{
Additional paramaters can be passed to \code{geom_jitter()} via \code{jitter.params}
and \code{\link[=theme]{theme()}} via \code{theme.params()}, where each \verb{\\*.params()} argument
takes a list
}
\examples{
herbarium_specimens \%>\%
  subset_coords(Longitude = c(-108, -105), Latitude = c(39, 42)) \%>\%
  dplyr::bind_cols(
    .,
    range_split(trait_tbl = ., split_var = "Mature_fruit_length_mm")
  ) \%>\%
  trait_phenology(
    specimen_tbl = .,
    trait = "Mature_fruit_length_mm_max",
    aesthetic_id = "Taxon_a_posteriori",
    legend_title = "Reviewed Annotations"
  )

}
\keyword{internal}
